% Generated by roxygen2 (4.0.2): do not edit by hand
\name{advanced_ds}
\alias{advanced_ds}
\title{A function to perform advanced divergence stratigraphy}
\usage{
advanced_ds(query_file, subject_set, eval = "1E-5", blast_path = NULL,
  blast_params = NULL, comp_cores = 1, dnds.threshold = 2,
  quiet = FALSE, clean_folders = FALSE)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_set}{a character string specifying the path to the CDS files of interest (subject organisms).}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{blast_path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{blast_params}{a character string listing the input paramters that shall be passed to the executing BLAST program. Default is \code{NULL}, implicating
that a set of default parameters is used when running BLAST.}

\item{comp_cores}{a numeric value specifying the number of cores that shall be used to perform
parallel computations on a multicore machine.}

\item{dnds.threshold}{a numeric value specifying the dnds threshold for genes that shall be retained.
Hence all genes having a dNdS value <= \code{dnds.threshold} are retained. Default is \code{dnds.threshold} = 2.}

\item{quiet}{a logical value specifying whether a successful interface call shall be printed out to the console.}

\item{clean_folders}{a logical value specifying whether the internal folder structure shall be deleted (cleaned) after
processing this function. Default is \code{clean_folders} = \code{FALSE}.}
}
\description{
This function uses an advanced methodology to perform divergence stratigraphy.
This new methodology needs more computation time, but aims to provide a more
accurate and more conservative pipeline for dNdS estimation and divergence stratum assignment.
}
\details{
This function ...
}
\author{
Hajk-Georg Drost
}

