% Generated by roxygen2 (4.0.2): do not edit by hand
\name{advanced_makedb}
\alias{advanced_makedb}
\title{Advanced interface function to makeblastdb}
\usage{
advanced_makedb(database_file, params, folder = "_blast_db/", path = NULL)
}
\arguments{
\item{database_file}{a character string specifying the path to the input file
that shall be transformed to a blast-able database.}

\item{params}{a character string specifying the arguments in the same notation as
calling makeblastdb from a shell like environment that shall be handed to the 'makeblastdb' call.
Examples could be: \code{params} = "-input_type fasta -dbtype prot -hash_index".}

\item{path}{a character string specifying the path to the makeblastdb program (in case you don't use the default path).
Default is \code{path} = \code{NULL}.}
}
\description{
This function provides a simple, but powerful interface
between the R language and 'makeblastdb'.
}
\examples{
\dontrun{

# make the A. thaliana genome to a blast-able database
advanced_makedb( database_file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
                 params = "-input_type fasta -dbtype prot -hash_index" )


}
}
\author{
Hajk-Georg Drost
}
\references{
Altschul, S.F., Gish, W., Miller, W., Myers, E.W. & Lipman, D.J. (1990) "Basic local alignment search tool." J. Mol. Biol. 215:403-410.

http://blast.ncbi.nlm.nih.gov/Blast.cgi?CMD=Web&PAGE_TYPE=BlastDocs&DOC_TYPE=ProgSelectionGuide

http://blast.ncbi.nlm.nih.gov/Blast.cgi
}

