% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alignmentSearch}
\alias{alignmentSearch}
\title{Orthologous Gene Detection based on best Global or Local Alignments}
\usage{
alignmentSearch(query_file, subject_file, seq_type = "prot",
  format = "fasta", tool = "ggsearch", path = NULL, comp_cores = 1,
  details = FALSE, statistics = FALSE, clean_folders = FALSE,
  quiet = TRUE)
}
\arguments{
\item{query_file}{a character string specifying the path to the sequence file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the sequence file of interest (subject organism).}

\item{seq_type}{a character string specifying the sequence type stored in the input file.
Options are are: "cds", "protein". In case of "cds", sequences are translated to protein sequences.
Default is \code{seq_type} = "protein".}

\item{format}{a character string specifying the file format of the sequence file. Default is \code{format} = \code{"fasta"}.}

\item{tool}{a character string specifying the global or local alignment tool that shall be used, e.g. \code{tool} = \code{"ggsearch"} or
\code{tool} = \code{"ssearch"}.}

\item{path}{a character string specifying the path to the tool executable (in case you don't use the default path).}

\item{comp_cores}{a numeric value specifying the number of cores to be used to perform best alignment searches.}

\item{details}{a boolean value specifying whether or not a detailed data.table shall be returned.}

\item{statistics}{a boolean value specifying whether the alignment statistics should be printed to the console and
shall also appear in the resulting data.table.}

\item{clean_folders}{a boolean value spefiying whether all internall folders storing the output of used programs
shall be removed. Default is \code{clean_folders} = \code{FALSE}.}

\item{quiet}{a logical value specifying whether or not the output of the alignment tool shall be printed to the console.}
}
\value{
a data.table containing query id and subject id as well as a series of score values.
}
\description{
This function performs a protein alignment for each sequence
from provided in the query_file to each of sequence of the subject_file and returns
a data.table objct storing the ids of the best global or local alignment pair and the corresponding
score parameter.
}
\examples{
\dontrun{

# performing an best alignment search between two protein fastas using Needleman Wunsch (GGSearch)
alignmentSearch( query_file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
                 subject_file = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'),
                 tool = "ggsearch")

# performing a search between two cds fastas in multicore mode (GGSearch)
alignmentSearch( query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
                 subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
                 comp_cores = 2, seq_type = "cds", tool = "ggsearch")


# performing an best alignment search between two cds fastas using Smith Waterman (SSearch)
alignmentSearch( query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
                 subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
                 tool="ssearch",comp_cores=1, seq_type = "cds")


# receive more detailed output
alignmentSearch( query_file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
                 subject_file = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'),
                 tool = "ggsearch", details = TRUE, statistics = TRUE )
   }
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

