% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.nr.R
\name{blast.nr}
\alias{blast.nr}
\title{Perform a BLASTP search against NCBI nr}
\usage{
blast.nr(query_file, nr.path = NULL, seq_type = "protein",
  format = "fasta", eval = "1E-5", max.target.seqs = 500, path = NULL,
  blast_params = NULL, comp_cores = 1)
}
\arguments{
\item{query_file}{a character string specifying the path to the query file of interest (query organism).}

\item{nr.path}{path to the NCBI nr database folder.}

\item{seq_type}{a character string specifying the sequence type stored in the input file. 
Options are are: "cds", "protein", or "dna". 
In case of "cds", sequence are translated to protein sequences, in case of "dna", 
cds prediction is performed on the corresponding sequences which subsequently are 
translated to protein sequences. Default is \code{seq_type} = \code{"protein"}.}

\item{format}{a character string specifying the file format of the sequence file, 
e.g. \code{format} = \code{"fasta"}. Default is \code{format} = \code{"fasta"}.}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{max.target.seqs}{a numeric value specifying the number of aligned sequences to keep.}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{blast_params}{a character string listing the input paramters that shall be passed to the executing BLAST program. Default is NULL, implicating that a set of default parameters is used when running BLAST.}

\item{comp_cores}{a numeric value specifying the number of cores that shall be used to run BLAST searches.}
}
\description{
This function performs a BLASTP search of a given set of sequences NCBI nr.
}
\details{
This function provides users an R interface to perform BLASTP searches against the NCBI nr database.
}
\examples{

\dontrun{

# perform a blastp serach of A. thaliana genes against NCBI nr
   blast.nr(
            query_file      = system.file('seqs/aa_seqs.fasta', package = 'orthologr'),
            nr.path         = "path/to/nr/database/folder",
            seq_type        = "protein",
            max.target.seqs = 1, 
            comp_cores      = 2 )
            
}

}
\author{
Hajk-Georg Drost
}
\references{
\url{http://www.ncbi.nlm.nih.gov/books/NBK1763/}
}
\seealso{
\code{\link{delta.blast}}, \code{\link{blast_best}}, \code{\link{blast_rec}}, \code{\link{blast}}, \code{\link{set_blast}}, \code{\link{advanced_blast}}
}

