% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blast_best}
\alias{blast_best}
\title{Function to perform a BLAST best hit search}
\usage{
blast_best(A, B)
}
\arguments{
\item{A}{an object returned by the set() function.}

\item{B}{an object returned by the set() function having the makedb parameter set to TRUE.}
}
\value{
A data.table as returned by the blast() function, storing the geneids
of orthologous genes (best hit) in the first column and the amino acid sequences in the second column.
}
\description{
This function performs a blast+ search (best hit) of a given set of protein sequences against a given database.
}
\details{
Given a set of protein sequences A, a best hit blast search is being performed from A to database.
}
\examples{
\dontrun{
# performing gene orthology inference using the best hit (BH) method
blast_best( set("ortho_lyra_aa.fasta") , set("ortho_thal_aa.fasta", makedb = TRUE))
}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

