% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.blast.R
\name{delta.blast}
\alias{delta.blast}
\title{Perform a DELTA-BLAST Search}
\usage{
delta.blast(query_file, subject_file, cdd.path = NULL, seq_type = "cds",
  format = "fasta", eval = "1E-5", max.target.seqs = 500, path = NULL,
  blast_params = NULL, comp_cores = 1, makedb_type = "protein")
}
\arguments{
\item{query_file}{a character string specifying the path to the query file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the subject file of interest (subject organism).}

\item{cdd.path}{path to the cdd database folder.}

\item{seq_type}{a character string specifying the sequence type stored in the input file. 
Options are are: "cds", "protein", or "dna". 
In case of "cds", sequence are translated to protein sequences, in case of "dna", 
cds prediction is performed on the corresponding sequences which subsequently are 
translated to protein sequences. Default is \code{seq_type} = \code{"cds"}.}

\item{format}{a character string specifying the file format of the sequence file, 
e.g. \code{format} = \code{"fasta"}. Default is \code{format} = \code{"fasta"}.}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{max.target.seqs}{a numeric value specifying the number of aligned sequences to keep.}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{blast_params}{a character string listing the input paramters that shall be passed to the executing BLAST program. Default is NULL, implicating that a set of default parameters is used when running BLAST.}

\item{comp_cores}{a numeric value specifying the number of cores that shall be used to run BLAST searches.}

\item{makedb_type}{a character string specifying the sequence type stored in the BLAST database that
is generated using 'makeblastdb'. Options are: "protein" and "nucleotide". Default is \code{makedb_type} = \code{"protein"}.}
}
\description{
This function performs a DELTA-BLAST search of a given set of sequences against a given database
or against the cdd database itself.
}
\details{
This function provides an interface between \code{delta-blast} and R. Given a query file, a delta-blast
search is perfored either against a subject database or against the cdd database itself.
}
\examples{
\dontrun{

# perform a delta-blast serach between A. thaliana and A. lyrata genes
delta.blast(
            query_file   = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
            cdd.path     = "path/to/cdd/database/folder" )
            
# perform a delta-blast serach between A. thaliana and the cdd database
delta.blast(
            query_file   = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
            subject_file = "cdd_delta",
            cdd.path     = "path/to/cdd/database/folder" )
}
}
\author{
Hajk-Georg Drost
}
\references{
\url{http://www.ncbi.nlm.nih.gov/books/NBK1763/#CmdLineAppsManual.Performing_a_DELTABLAS}
}
\seealso{
\code{\link{blast_best}}, \code{\link{blast_rec}}, \code{\link{blast}}, \code{\link{set_blast}}, \code{\link{advanced_blast}}
}

