% Generated by roxygen2 (4.0.2): do not edit by hand
\name{divergence_stratigraphy}
\alias{divergence_stratigraphy}
\title{Function to perform divergence stratigraphy}
\usage{
divergence_stratigraphy(query_file, subject_file, eval = "1E-5",
  ortho_detection = "RBH", blast_path = NULL, mafft_path = NULL,
  comp_cores = 1, dnds.threshold = 2, quiet = FALSE,
  clean_folders = FALSE, ds.values = TRUE)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{ortho_detection}{a character string specifying the orthology inference method that shall be performed
to detect orthologous genes. Default is \code{ortho_detection} = "RBH" (BLAST reciprocal best hit).
Available methods are: "BH" (BLAST best hit), "RBH" (BLAST reciprocal best hit).}

\item{blast_path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{mafft_path}{a character string specifying the path to the multiple alignment program MAFFT (in case you don't use the default path).}

\item{comp_cores}{a numeric value specifying the number of cores that shall be used to perform
parallel computations on a multicore machine.}

\item{dnds.threshold}{a numeric value specifying the dnds threshold for genes that shall be retained.
Hence all genes having a dNdS value <= \code{dnds.threshold} are retained. Default is \code{dnds.threshold} = 2.}

\item{quiet}{a logical value specifying whether a successful interface call shall be printed out to the console.}

\item{clean_folders}{a logical value specifying whether the internal folder structure shall be deleted (cleaned) after
processing this function. Default is \code{clean_folders} = \code{FALSE}.}

\item{ds.values}{a logical value specifying whether divegrence stratum values (ds values) or dNdS values shall be returned
by \code{divergence_stratigraphy}. Default is \code{ds.values} = \code{TRUE}.}
}
\value{
A data.table storing the divergence map of the query organism.
}
\description{
This function takes a query organism and performs
divergence stratigraphy (Quint et al.,2012 ; Drost et al. 2014) against a
closely related subject organism.
}
\details{
Introduced by Quint et al., 2012 and extended in Drost et al. 2014, divergence stratigraphy
 is the process of quantifying the selection pressure (in terms of amino acid sequence divergence) acting on
 orthologous genes between closely related species. The resulting sequence divergence map (short divergence map),
 stores the divergence stratum in the first column and the query_id of inferred orthologous genes in the second column.

 Following steps are performed to obtain a standard divergence map based on divergence_stratigraphy:

 1) Orthology Inference using BLAST reciprocal best hit ("RBH")

 2) Pairwise amino acid alignments of orthologous genes using the MAFFT program

 3) Codon alignments of orthologous genes using PAL2NAL

 4) dNdS estimation using Comeron's method (1995)

 5) Assigning dNdS values to divergence strata (deciles)
}
\examples{
\dontrun{


 # performing standard divergence stratigraphy
 divergence_stratigraphy(
      query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
      subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
      eval = "1E-5", ortho_detection = "RBH", comp_cores = 1,
      quiet = TRUE, clean_folders = TRUE)



 # performing standard divergence stratigraphy using the mafft_path argument to specify
 # the path to the MAFFT program
 divergence_stratigraphy(
      query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
      subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
      eval = "1E-5", ortho_detection = "RBH",mafft_path = "path/to/mafft",
      comp_cores = 1, quiet = TRUE, clean_folders = TRUE)


 # Divergence Stratigraphy can also be performed in parallel
 # (on a multicore machine) using the 'comp_cores' argument
 divergence_stratigraphy(
      query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
      subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
      eval = "1E-5", ortho_detection = "RBH", comp_cores = 2,
      quiet = TRUE, clean_folders = TRUE)


 }
}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}
\references{
Quint M et al. (2012). "A transcriptomic hourglass in plant embryogenesis". Nature (490): 98-101.

 Drost HG et al. (2014). "Active maintenance of phylotranscriptomic hourglass patterns in animal and plant embryogenesis".
}
\seealso{
\code{\link{dNdS}}, \code{\link{substitutionrate}}, \code{\link{multi_aln}},
  \code{\link{codon_aln}}, \code{\link{blast_best}}, \code{\link{blast_rec}}
}

