% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exonerate.R
\name{exonerate}
\alias{exonerate}
\title{Sequence Comparison with Exonerate}
\usage{
exonerate(query_files, subject_files, params = NULL, core_interface = FALSE,
  format = "fasta", clean_folders = TRUE)
}
\arguments{
\item{query_files}{a character vector specifying the paths to the fasta files of interest (query organism(s)).}

\item{subject_files}{a character vector specifying the paths to the fasta files of interest (subject organism(s)).}

\item{params}{a character string listing the input paramters that shall be passed to the executing Exonerate program.
Default is \code{params} = \code{NULL}, implicating that a set of default parameters is used when running Exonerate.}

\item{core_interface}{a logical value enabling full flexibility to call exonerate with all additional parameters.
When \code{core_interface} = \code{TRUE}, only \code{system("exonerate ",params)} is being called and you can specify any
parameter constillation using the \code{params} argument. Note that you also have to specify the paths to you query and subject fasta files
when \code{core_interface} = \code{TRUE}. See examples for details.}

\item{format}{a character string specifying the file format of the sequence file. Right now only the 'fasta' format is supported.}

\item{clean_folders}{a boolean value specifying whether the internall "_exonerate" folder storing the output of exonerate shall be removed.
Default is \code{clean_folders} = \code{TRUE}.}
}
\value{
a sequence alignment as returned by the Exonerate program.
}
\description{
This function provides an interface to the Exonerate program: a generic tool for sequence alignment.
}
\details{
When using this interface function it is assumed that you have the Exonerate program installed on your machine
and it can be executed from the default \emph{PATH} environment.

See \url{https://www.ebi.ac.uk/~guy/exonerate/exonerate.man.html} for detailed parameter specifications when using the Exonerate program.


Internally this function creates a folder "_exonerate" to store the corresponding output of
the Exonerate programs and will delete it (in case \code{clean_folders} = \code{TRUE}) after the output is
stored as R seqinr alignment object.

In case you want to retain the output of Exonerate, please select \code{clean_folders} = \code{FALSE}.
}
\examples{
\dontrun{

# a simple ungapped alignment returned by exonerate
exonerate(query_file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
          subject_file = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'))


### peforming a Exonerate alignment using additional parameters
# 6-frame translated alignment: param = "--model coding2coding"
exonerate(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
          subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
          params = "--model coding2coding")


### peforming a Exonerate with full flexibility
exonerate(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
          subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
          params = paste0("--model coding2coding",
          system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
          " ",system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
          core_interface = TRUE)

}
}
\author{
Hajk-Georg Drost
}
\references{
Guy St C Slater and Ewan Birney (2005). Automated generation of heuristics for biological sequence comparison. BMC Bioinformatics 6:31.

\url{https://www.ebi.ac.uk/~guy/exonerate/index.html}
}

