% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_best_hit.R
\name{filter_best_hits}
\alias{filter_best_hits}
\title{Helper function to select best BLAST hit based on minimum evalue}
\usage{
filter_best_hits(x)
}
\arguments{
\item{x}{a tibble storing gene locus ids, splice variant ids, and blast output for filtering}
}
\description{
Helper function to select best BLAST hit based on minimum evalue
}
\details{
Extracts the best hit based on lowest e-value and longest splice variant 
when e-values are equal
}
\author{
Hajk-Georg Drost
}
