% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ortholog_tables.R
\name{generate_ortholog_tables}
\alias{generate_ortholog_tables}
\title{Generate ortholog tables by gene locus and splice varaint}
\usage{
generate_ortholog_tables(dNdS_file, annotation_file_query,
  annotation_file_subject, format)
}
\arguments{
\item{dNdS_file}{file path to a \code{dNdS_tbl} generated with \code{\link{dNdS}} and
stored conform with \code{\link{read.dnds.tbl}}.}

\item{annotation_file_query}{file path to the annotation file of the query species in \code{gtf} or \code{gff} file format.}

\item{annotation_file_subject}{file path to the annotation file of the subject species in \code{gtf} or \code{gff} file format.}

\item{format}{a vector of length 2 storing the annotation file formats of the query annotation file and subject annotation file: either \code{gtf} or \code{gff} format. E.g. \code{format = c("gtf","gtf")}.}
}
\description{
Given an input dNdS table generated by \code{\link{dNdS}} and
annotation files for the query and subject species in \code{gtf} or \code{gff}
file format, this function selects the best BLAST hit to represent either a gene locus (e.g. the splice variant of the gene locus with lowest e-value) or the best BLAST hit for a splice varaint.
}
\author{
Hajk-Georg Drost
}
