% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggsearch}
\alias{ggsearch}
\title{Perform Global Alignments with GGSearch}
\usage{
ggsearch(file, library_file, options = "", lib_format = "fasta",
  seq_type = "protein", path = NULL, parse_output_to = NULL)
}
\arguments{
\item{file}{a character string specifying the path to the query sequence file.}

\item{library_file}{a character string specifying the path to the subject library file.}

\item{options}{a character string specifying additional options for GGSearch.}

\item{lib_format}{a character string specifying the file type of the library e.g. \code{lib_format} = \code{"fasta"}.}

\item{seq_type}{a character string specifying the type of biological sequences.}

\item{path}{a character string specifying the path to GGSearch. This argument only needs
to be specified when GGSearch cannot be executed from the default execution \code{PATH} (global environment).}

\item{parse_output_to}{a charcater string specifying a filename in which the output of
GGSearch is written.}
}
\description{
This function allows you to perform global alignments between pairs
of protein sequences or alignments between a protein sequence against a
protein database based on the GGSeach program implemented in FASTA framework.
}
\details{
This function takes two fasta files representing query and subject (library) sequences.

As a result, this function returns the alignment score and additional parameters
computed by GGSearch. GGSearch implements a Needleman-Wunsch like alignment, except that affine gap
penalties are used.
}
\examples{
\dontrun{

# performing GGSearch on two protein files
ggsearch( file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
          library_file = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'))

# performing GGSearch on two cds files parsing the output
ggsearch( file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
          library_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
          seqtype = "cds", parse_output_to = "ggsearch.out")

}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}
\references{
GGSearch is part of the fasta36 package that can be downloaded here

\url{http://fasta.bioch.virginia.edu/fasta_www2/fasta_down.shtml}
}

