% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ortholog_tables_all.R
\name{import_ortholog_tables_all}
\alias{import_ortholog_tables_all}
\title{Importing output pairwise orthologs tables generated with \code{\link{generate_ortholog_tables_all}}}
\usage{
import_ortholog_tables_all(ortholog_tables_folder)
}
\arguments{
\item{ortholog_tables_folder}{file path to output files generated with \code{\link{generate_ortholog_tables_all}}.}
}
\description{
Importing output pairwise orthologs tables generated with \code{\link{generate_ortholog_tables_all}}.
}
\author{
Hajk-Georg Drost
}
