% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnc_map_core_orthologs.R
\name{lnc_map_core_orthologs}
\alias{lnc_map_core_orthologs}
\title{Retrieve a core set of orthologous lncRNAs (deprecated)}
\usage{
lnc_map_core_orthologs(lnc_map, species_order)
}
\arguments{
\item{lnc_map}{a \code{lnc_map} that was generated with \code{\link{map_generator_lnc}}.}

\item{species_order}{a character string specifying species names listed in the order of
phylogenetic/taxonomic distance from the query species.}
}
\description{
This function is deprecated and will be removed in a future release.
Please use \code{\link{retrieve_core_orthologs}} instead, which now handles both
protein-coding gene tables and lncRNA maps automatically.
}
\author{
Hajk-Georg Drost
}
