% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnc_map_counts.R
\name{lnc_map_counts}
\alias{lnc_map_counts}
\title{Count number of orthologous lncRNAs per pairwise species comparison}
\usage{
lnc_map_counts(lnc_map, species_order)
}
\arguments{
\item{lnc_map}{a orthologous lncRNA map generated with \code{\link{map_generator_lnc}}.}

\item{species_order}{character string specifying species names listed in the order of phylogenetic/taxonomic distance from the query species.
The species names must match with the species names present in \code{map_generator_lnc}.}
}
\description{
This function takes a orthologous lncRNA map generated with \code{\link{map_generator_lnc}}
as input and groups the table by species to count the number of orthologous lncRNAs per pairwise species comparison
using the reference (query) species specified in \code{\link{map_generator_lnc}}.
}
\author{
Hajk-Georg Drost
}
