% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generator.lnc.R
\name{map.generator.lnc}
\alias{map.generator.lnc}
\title{Infer orthologous lncRNAs between muliple species}
\usage{
map.generator.lnc(query_file, subjects.folder, output.folder,
  eval = "1E-5", ortho_detection = "RBH", comp_cores = 1,
  progress.bar = TRUE, sep = ";", ...)
}
\arguments{
\item{query_file}{a character string specifying the path to the lncRNAs file of the query organism in \code{fasta} format.}

\item{subjects.folder}{a character string specifying the path to the folder where lncRNAs files in \code{fasta} format of the subject organisms are stored.}

\item{output.folder}{a character string specifying the path to the folder where output orthologous tables should be stored.}

\item{eval}{a character string specifying the e-value for BLAST based orthology inference. Please use the scientific notation.}

\item{ortho_detection}{a character string specifying the Orthology Inference method that shall be used to perform
dNdS computations. Possible options are: 
\itemize{
 \item \code{ortho_detection = "BH"}: BLAST best unidirectional hit
 \item \code{ortho_detection = "RBH"}: BLAST best reciprocal hit
}}

\item{comp_cores}{number of computing cores that shall be used to perform parallelized computations.}

\item{progress.bar}{should a progress bar be shown. Default is \code{progress.bar = TRUE}.}

\item{sep}{a file separator that is used to store maps as csv file.}

\item{...}{additional parameters that shall be passed to  \code{\link{dNdS}}.}
}
\description{
Inference of orthologous lncRNAs between multiple species is performed via pairwise comparisons.
 The corresponding orthologous tables are then stored in an output folder.
}
\details{
Given a query organism and a set of subject organsisms that are stored in the same folder,
this function crawls through all subject organsism and infers the lncRNA homologs in
pairwise species comparisons.
}
\examples{
\dontrun{
map.generator.lnc(
   query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
   subjects.folder = system.file('seqs/map_gen_example', package = 'orthologr'),
   output.folder   = getwd(),
   quiet           = TRUE,
   comp_cores      = 1
)
}
}
\author{
Hajk-Georg Drost
}
