% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generator.R
\name{map_generator_dnds}
\alias{map_generator_dnds}
\title{Generate dNdS Maps Between a Query Organism and Multiple Subject Organisms}
\usage{
map_generator_dnds(
  query_file,
  subjects_folder,
  output_folder,
  eval = "1E-5",
  min_qry_coverage_hsp = 50,
  min_qry_perc_identity = 30,
  ortho_detection = "RBH",
  aa_aln_type = "pairwise",
  aa_aln_tool = "NW",
  codon_aln_tool = "pal2nal",
  dnds_est_method = "Comeron",
  comp_cores = 1,
  progress_bar = TRUE,
  sep = ";",
  ...
)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of the query organism.}

\item{subjects_folder}{a character string specifying the path to the folder where CDS files of the subject organisms are stored.}

\item{output_folder}{a character string specifying the path to the folder where output dNdS/DS Maps should be stored.}

\item{eval}{a character string specifying the e-value for BLAST based Orthology Inference that is performed
in the process of dNdS computations. Please use the scientific notation.}

\item{min_qry_coverage_hsp}{minimum \code{qcovhsp} (= query coverage of the HSP) of an orthologous hit (a value between 1 and 100).}

\item{min_qry_perc_identity}{minimum \code{perc_identity} (= percent sequence identity between query and selected HSP) of an orthologous hit (a value between 1 and 100).}

\item{ortho_detection}{a character string specifying the Orthology Inference method that shall be used to perform
dNdS computations. Possible options are: \code{ortho_detection} = \code{"BH"} (BLAST best hit), 
\code{ortho_detection} = \code{"RBH"} (BLAST best reciprocal hit), etc.}

\item{aa_aln_type}{a character string specifying the amino acid alignement type: \code{aa_aln_type = "multiple"} or \code{aa_aln_type = "pairwise"}. 
Default is \code{aa_aln_type = "pairwise"}.}

\item{aa_aln_tool}{a character string specifying the program that should be used e.g. "clustalw".}

\item{codon_aln_tool}{a character string specifying the codon alignment tool that shall be used. 
Default is \code{codon_aln_tool = "pal2nal"}. Right now only "pal2nal" can be selected as codon alignment tool.}

\item{dnds_est_method}{a character string specifying the dNdS estimation method, e.g. "Comeron","Li", "YN", etc. See Details for all options.}

\item{comp_cores}{number of computing cores that shall be used to perform parallelized computations.}

\item{progress_bar}{should a progress bar be shown. Default is \code{progress_bar = TRUE}.}

\item{sep}{a file separator that is used to store maps as csv file.}

\item{...}{additional parameters that shall be passed to  \code{\link{dNdS}}.}
}
\description{
This function allows you to compute dNdS or DS Maps between a query organism
and a set subject organisms stored in the same folder. The corresponding dNdS/DS Maps are then stored in an output folder.
}
\details{
Given a query organism and a set of subject organsisms that are stored in the same folder,
this function crawls through all subject organsism and as a first step computes the pairwise 
dNdS Maps between query and subject organsim and as a second step stores the corresponding Map 
in an output folder.
}
\examples{
\dontrun{
map_generator_dnds(
   query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
   subjects_folder = system.file('seqs/map_gen_example', package = 'orthologr'),
   aa_aln_type      = "pairwise",
   aa_aln_tool      = "NW", 
   codon_aln_tool   = "pal2nal", 
   dnds_est_method  = "Comeron",
   output_folder   = "orthologr_dnds_maps",
   quiet           = TRUE,
   comp_cores      = 1
)

}
}
\author{
Hajk-Georg Drost
}
