% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OF2CoreOrthologs.R
\name{OF2CoreOrthologs}
\alias{OF2CoreOrthologs}
\title{Retrieve core orthologs across multiple species from Orthofinder2 output}
\usage{
OF2CoreOrthologs(orthogroups_file, single_copy_file)
}
\arguments{
\item{orthogroups_file}{file path to Orthofinder2 output file named \code{Orthogroups.csv}.}

\item{single_copy_file}{file path to Orthofinder2 output file named \code{SingleCopyOrthogroups.txt}.}
}
\description{
Retrieve core orthologs (= single copy orthologs) across multiple species
from a Orthofinder2 output.
}
\author{
Hajk-Georg Drost
}
