% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ortho_lyra_aa.fasta}
\alias{ortho_lyra_aa.fasta}
\title{Fasta file storing 20 hand curated protein sequences of
A lyrata genes.}
\format{A fasta file storing gene ids and amino acid sequences of 20 A. lyrata genes.}
\description{
A dataset containing 20 hand curated protein sequences of
A lyrata genes stored in a fasta file. This dataset can be
used for a small test of the orthologr pipeline and shall respresent a small
sample proteome of A. lyrata.
}

