\name{orthologr-package}
\alias{orthologr-package}
\alias{orthologr}
\docType{package}
\title{
A package combining common gene orthology detection methods and dN/dS models for pipeline processing.
}
\description{
This package aims to provide a toolbox and functionality to perform gene orthology detection using state of the art software tools and dN/dS computations for evoutionary and population genetics studies.
}
\details{
\tabular{ll}{
Package: \tab orthologr\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-08-23\cr
License: \tab MIT\cr
}


}
\author{
Hajk-Georg Drost, Kristian Karsten Ullrich, Sarah Scharfenberg, Norman Urban

Maintainer: Hajk-Georg Drost <hajk-georg.drost@informatik.uni-halle.de> and Kristian Karsten Ullrich
}
\references{

}
\keyword{ package }

\examples{
%% ~~ simple examples of the most important functions ~~
}
