% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/OrthoMCL.R
\name{OrthoMCL}
\alias{OrthoMCL}
\title{Orthology Inference with OrthoMCL}
\usage{
OrthoMCL(query_file, subject_files, orthomcl_params = NULL, eval = "1E-5",
  seq_type = "protein", format = "fasta", orthomcl_path = NULL,
  comp_cores = 1, delete_files = FALSE)
}
\arguments{
\item{query_file}{a character string specifying the path to the sequence file of interest (query organism).}

\item{subject_files}{a character string specifying the paths to the sequence files of interest (subject organisms).}

\item{orthomcl_params}{a character string specifying additional parameters that shall be handed to the OrthoMCL call,
e.g. \code{orthomcl_params} = \code{""}. See \url{http://www.orthomcl.org/orthomcl/about.do}
for parameter details.}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{seq_type}{a character string specifying the sequence type stored in the input file.
Options are: "protein", or "dna". OrthoMCL can only handle protein or nucleotide sequences stored in fasta files.}

\item{format}{when using OrthoMCL you can only specify \code{format} = \code{"fasta"}.}

\item{orthomcl_path}{a character string specifying the execution path to OrthoMCL.}

\item{comp_cores}{a numeric value specifying the number of cores that shall be
used to run OrthoMCL.}

\item{delete_files}{a boolean value specifying whether the folder '_OrthoMCL' that stored the
OrthoMCL output files shall be removed after the analysis. Default is \code{delete_files} = \code{FALSE}.}
}
\value{
a data.frame storing the OrthoMCL output.
}
\description{
This function takes a query organism and a set of subject organisms
and performs orthology inference using OrthoMCL as orthology detection program (methodology).
}
\details{
This function provides an interface to the prthology detection program OrthoMCL.

To use this function \emph{OrthoMCL} needs to be installes and executable on your system.
}
\examples{
\dontrun{

# finding orthologs between Arabidopsis thaliana and Arabidopsis lyrata genes
orthologs(query_file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
          subject_files = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'),
          seq_type = "protein", ortho_detection = "OrthoMCL")


}
}
\author{
Hajk-Georg Drost
}
\references{
Li Li, Christian J. Stoeckert, Jr., and David S. Roos
OrthoMCL: Identification of Ortholog Groups for Eukaryotic Genomes
Genome Res. 2003 13: 2178-2189.

Feng Chen, Aaron J. Mackey, Christian J. Stoeckert, Jr., and David S. Roos
OrthoMCL-DB: querying a comprehensive multi-species collection of ortholog groups
Nucleic Acids Res. 2006 34: D363-8.

Feng Chen, Aaron J. Mackey, Jeroen K. Vermunt, and David S. Roos
Assessing Performance of Orthology Detection Strategies Applied to Eukaryotic Genomes
PLoS ONE 2007 2(4): e383.

Fischer, S., Brunk, B. P., Chen, F., Gao, X., Harb, O. S., Iodice, J. B., Shanmugam, D., Roos, D. S. and Stoeckert, C. J.
Using OrthoMCL to Assign Proteins to OrthoMCL-DB Groups or to Cluster Proteomes Into New Ortholog Groups
Current Protocols in Bioinformatics. 2011 35:6.12.1-6.12.19.

\url{http://www.orthomcl.org/orthomcl/}
}
\seealso{
\code{\link{orthologs}}, \code{\link{dNdS}}
}

