% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotor_divergence_estimation.R
\name{promotor_divergence_estimation}
\alias{promotor_divergence_estimation}
\title{Estimate the DNA distance between promotor sequences}
\usage{
promotor_divergence_estimation(query, subject, model = "K80")
}
\arguments{
\item{query}{file path to a query \code{fasta} file storing promotor sequences of interest (e.g. generated with \code{\link[metablastr]{extract_upstream_promotor_seqs}}).}

\item{subject}{file path to a subject \code{fasta} file storing promotor sequences of interest (e.g. generated with \code{\link[metablastr]{extract_upstream_promotor_seqs}}).}

\item{model}{a model as specified in \code{\link[ape]{dist.dna}}: a character string specifying the evolutionary model to be used - must be one of:
\itemize{
\item  \code{K80} (the default)
\item \code{raw}
\item  \code{N}
\item  \code{TS}
\item  \code{TV}
\item  \code{JC69}
\item  \code{F81} 
\item \code{K81}
\item \code{F84}
\item \code{BH87}
\item \code{T92}
\item \code{TN93}
\item \code{GG95}
\item \code{logdet}
\item \code{paralin}
}}
}
\description{
Given two \code{fasta} files storing the promotor sequence
of protein coding genes, this function estimates the pairwise DNA distance between these promotor sequences.
}
\author{
Hajk-Georg Drost
}
