% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promotor_divergence_of_orthologous_genes.R
\name{promotor_divergence_of_orthologous_genes}
\alias{promotor_divergence_of_orthologous_genes}
\title{Compute promotor sequence divergence of orthologous genes}
\usage{
promotor_divergence_of_orthologous_genes(
  promotor_folder,
  ortholog_tables_folder,
  model = "K80",
  ortholog_promotor_seq_output = NULL
)
}
\arguments{
\item{promotor_folder}{a path to a folder containing promotor sequences in \code{fasta} format that
were generated with \code{\link[metablastr]{extract_upstream_promotor_seqs}}.}

\item{ortholog_tables_folder}{a path to a folder containing dNdS tables generated with \code{\link{generate_ortholog_tables_all}}}

\item{model}{a model as specified in \code{\link[ape]{dist.dna}}: a character string specifying the evolutionary model to be used - must be one of:
\itemize{
\item \code{K80} (the default)
\item \code{raw}
\item \code{N}
\item \code{TS}
\item \code{TV}
\item \code{JC69}
\item \code{F81} 
\item \code{K81}
\item \code{F84}
\item \code{BH87}
\item \code{T92}
\item \code{TN93}
\item \code{GG95}
\item \code{logdet}
\item \code{paralin}
}}

\item{ortholog_promotor_seq_output}{a path or name to an output folder where orthologous promotor sequences shall be stored.}
}
\description{
This function computes the promotor sequence divergences of orthologous genes 
from a set of pairwise species comparisons. It allows to add this promotor divergence information
to an pre-computed \code{dNdS table} generated with \code{\link{dNdS}} or \code{\link{generate_ortholog_tables_all}}.
}
\author{
Hajk-Georg Drost
}
