% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dnds.tbl.R
\name{read.dnds.tbl}
\alias{read.dnds.tbl}
\title{Import a dNdS table generated with \code{dNdS}}
\usage{
read.dnds.tbl(file)
}
\arguments{
\item{file}{file path to dNdS table with \code{;} separated columns.}
}
\description{
This function reads a file that stores a dnds table
generated with \code{\link{dNdS}}.
}
\examples{
# generate dNdS table
dNdS_tbl <- dNdS(query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
subject_file    = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
ortho_detection = "RBH", 
aa_aln_type     = "pairwise",
aa_aln_tool     = "NW", 
codon_aln_tool  = "pal2nal", 
dnds_est.method = "Comeron", 
comp_cores      = 1 )

# save dNdS table as ';' column separated file
utils::write.table(
dNdS_tbl, 
file.path(tempdir(), "dNdS_tbl.csv"), sep = ";", 
col.names = TRUE,
row.names = FALSE,
quote     = FALSE )

# import dNdS table into R session
dNdS_tbl_import <- read.dnds.tbl(file.path(tempdir(), "dNdS_tbl.csv"))
}
\author{
Hajk-Georg Drost
}
