% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read.proteome}
\alias{read.proteome}
\title{Read the proteome of a given organism}
\usage{
read.proteome(file, format, ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the proteome.}

\item{format}{a character string specifying the file format used to store the proteome, e.g. "fasta", "fastq".}

\item{...}{additional arguments that are used by the \code{\link[Biostrings]{readAAStringSet}} function.}
}
\value{
A data.table storing the gene id in the first column and the corresponding
sequence as string in the second column.
}
\description{
This function reads an organism specific proteome stored in a defined file format.
}
\details{
The \code{read.proteome} function takes a string specifying the path to the proteome file
of interest as first argument.

It is possible to read in different proteome file standards such as \emph{fasta} or \emph{fastq}.

Proteomes stored in fasta files can be downloaded from https://www.ebi.ac.uk/reference_proteomes, etc.
}
\examples{
\dontrun{
# reading a proteome stored in a fasta file
Ath.proteome <- read.proteome(system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
                               format = "fasta")
}

}
\author{
Hajk-Georg Drost
}
