% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_longest_isoforms.R
\name{retrieve_longest_isoforms}
\alias{retrieve_longest_isoforms}
\title{Retrieve the longest isoforms from a proteome file and save results as fasta file}
\usage{
retrieve_longest_isoforms(proteome_file, gtf_file, new_file)
}
\arguments{
\item{proteome_file}{file path to proteome in \code{fasta} file format.}

\item{gtf_file}{file path to the corresponding annotation file in \code{gtf} file format.}

\item{new_file}{file path to new file storing only peptide sequences of the longest isoforms.}
}
\description{
Based on a fasta file storing the peptide isoforms of gene loci and
an annotation file in \code{gtf} file format, this function extracts the longest
isoform per gene locus and stores the results in a new \code{fasta} file.
This procedure enables easier downstream analyses such as orthology inference etc
when dealing with proteome \code{fasta} files which usually include isoform peptides.
}
\examples{
\dontrun{
# retrieve example data from ENSEMBLGENOMES
proteome <- biomartr::getProteome(db = "ensemblgenomes", organism = "Arabidopsis thaliana")
biomartr::getGTF(db = "ensemblgenomes", organism = "Arabidopsis thaliana")
R.utils::gunzip("ensembl/annotation/Arabidopsis_thaliana.TAIR10.42_ensemblgenomes.gtf.gz")
annotation <- "ensembl/annotation/Arabidopsis_thaliana.TAIR10.42_ensemblgenomes.gtf"
# retrieve longest isoforms and store in new file
retrieve_longest_isoforms(proteome_file = proteome, 
                          gtf_file = annotation, 
                          new_file = "Athaliana_pep_longest.fa")
# import new file into R session                          
Athaliana_pep_longest <- Biostrings::readAAStringSet("Athaliana_pep_longest.fa")
}
}
\author{
Hajk-Georg Drost
}
