% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_longest_isoforms_all.R
\name{retrieve_longest_isoforms_all}
\alias{retrieve_longest_isoforms_all}
\title{Retrieve the longest isoforms of several proteome files stored in a folder}
\usage{
retrieve_longest_isoforms_all(proteome_folder, annotation_folder,
  output_folder, annotation_format = "gff")
}
\arguments{
\item{proteome_folder}{file path to proteome in \code{fasta} file format.}

\item{annotation_folder}{file path to the corresponding annotation file in \code{gtf} file format.}

\item{output_folder}{file path to new file storing only peptide sequences of the longest isoforms.}

\item{annotation_format}{format of \code{annotation_file}. Options are:
\itemize{
\item \code{annotation_file = "gff"} (default)
\item \code{annotation_file = "gtf"}
}}
}
\description{
Based on a fasta file storing the peptide isoforms of gene loci and
an annotation file in \code{gtf} file format, this function extracts the longest
isoform per gene locus and stores the results in a new \code{fasta} file.
This procedure enables easier downstream analyses such as orthology inference etc
when dealing with proteome \code{fasta} files which usually include isoform peptides.
}
\examples{
\dontrun{
orgs <- c("Arabidopsis lyrata", 
          "Capsella rubella", "Solanum lycopersicum")
# download proteome files for all species          
biomartr::getProteomeSet(db = "refseq", organisms = orgs, path = "of_proteomes")
# download annotation files for all species          
biomartr::getGFFSet(db = "refseq", organisms = orgs, path = "of_gff")
# select longest splice variant per gene locus
retrieve_longest_isoforms_all(proteome_folder = "of_proteomes", 
                              annotation_folder = "of_gff",
                              annotation_format = "gff", 
                              output_folder = "of_proteomes_longest_sv")
}
}
\author{
Hajk-Georg Drost
}
