% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_orthologs.R
\name{select_orthologs}
\alias{select_orthologs}
\title{Select orthologs based on either gene locus or splice variant}
\usage{
select_orthologs(
  dnds_tbl,
  annotation_file_query,
  annotation_file_subject,
  collapse_by = "gene_locus",
  format = c("gtf", "gtf")
)
}
\arguments{
\item{dnds_tbl}{a tibble returned by \code{\link{dNdS}}.}

\item{annotation_file_query}{file path to query annotation file in either \code{gtf} or \code{gff} format.}

\item{annotation_file_subject}{file path to query annotation file in either \code{gtf} or \code{gff} format.}

\item{collapse_by}{locus type by which orthologs should be determined. Options are:
\itemize{
\item \code{type = "gene_locus"}: the splice variant having the lowest e-value of the BLAST search is chosen to represent the gene locus for the orthology relationship.
\item \code{type = "splice_variant"}: all splice variants of homologous gene loci will be returned.
}}

\item{format}{a vector of length 2 storing the annotation file formats of the query annotation file and subject annotation file: either \code{gtf} or \code{gff} format. E.g. \code{format = c("gtf","gtf")}.}
}
\description{
This function selects orthologs based on either gene locus or splice variant.
}
\author{
Hajk-Georg Drost
}
