% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_blast}
\alias{set_blast}
\title{Function preparing the parameters and databases for subsequent BLAST searches}
\usage{
set_blast(file, seq_type = "cds", format = "fasta", makedb = FALSE,
  path = NULL, makedb_type = "protein", ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the sequences of interest.}

\item{seq_type}{a character string specifying the sequence type stored in the input file.
Options are are: "cds", "protein", or "dna". In case of "cds", sequence are translated to protein sequences,
in case of "dna", cds prediction is performed on the corresponding sequences which subsequently are
translated to protein sequences. Default is \code{seq_type} = "cds".}

\item{format}{a character string specifying the file format used to store the genome, e.g. "fasta", "gbk".}

\item{makedb}{TRUE or FALSE whether a database should be created or not (BLAST parameter 'makeblastdb').}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{makedb_type}{a character string specifying the sequence type stored in the BLAST database
that is generated using 'makeblastdb'. Options are: "protein" and "nucleotide". Default is \code{makedb_type} = "protein".}

\item{...}{additional arguments that are used by the seqinr::read.fasta() function.}
}
\value{
A list storing two elements. The first element [[1]] corresponds to the data.table storing the gene ids in the first column and
the corresponding dna (cds) sequence in the second column and the aminoacid sequence third column.
The second list element [[2]] stores the name of the protein database that was created by 'makeblastdb'.
}
\description{
This function reads a file storing a specific sequence type, such as "cds", "protein", or
"dna" in a standard sequence file format such as "fasta", etc. and depending of the \code{makedb}
 parameter either creates a blast-able database, or returns the corresponding protein sequences
 as data.table object for further BLAST searches.
}
\examples{
\dontrun{
 # running the set function to see an example output
 head(set_blast(file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'))[[1]] , 2)
}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}
\seealso{
\code{\link{blast_best}}, \code{\link{blast_rec}}, \code{\link{advanced_blast}}, \code{\link{blast}}
}

