% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ssearch}
\alias{ssearch}
\title{Compute Local Alignments with SSearch}
\usage{
ssearch(file, library_file, options = "", lib_format = "fasta",
  seq_type = "protein", path = NULL, parse_output_to = NULL)
}
\arguments{
\item{file}{a character string specifying the path to the query sequence file.}

\item{library_file}{a character string specifying the path to the subject library file.}

\item{options}{a character string specifying additional options for SSearch.}

\item{lib_format}{a character string specifying the file type of the library e.g. \code{lib_format} = \code{"fasta"}.}

\item{seq_type}{a character string specifying the type of biological sequences.}

\item{path}{a character string specifying the path to ssearch if not included in
the global environment.}

\item{parse_output_to}{a charcater string specifying a filename in which the output of
SSearch is written.}
}
\description{
This function allows you to perform local alignments between pairs
of protein sequences or alignments between a protein sequence against a
protein database based on the SSeach program implemented in FASTA framework.
}
\details{
This function takes two fasta files representing query and subject (library) sequences and computes
Local Pairwise Alignments.

As a result, this function returns the alignment score and additional parameters
computed by SSearch. SSearch performs a rigorous Smith-Waterman alignment between protein sequences.
}
\examples{
\dontrun{

# performing SSearch on two protein files
ssearch( file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
          library_file = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'))

# performing SSearch on two cds files parsing the output
ssearch( file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
          library_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
          seq_type = "cds", parse_output_to = "ssearch.out")

}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}
\references{
SSearch is part of the fasta36 package that can be found here

\url{http://fasta.bioch.virginia.edu/fasta_www2/fasta_down.shtml}

Pearson WR and Lipman DJ. 1988. Improved tools for biological sequence comparison. PNAS 85: 2444-2448.

\url{http://www.pnas.org/content/85/8/2444.long}
}

