% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCoreOrthoParamsGeneLocus.R
\name{testCoreOrthoParamsGeneLocus}
\alias{testCoreOrthoParamsGeneLocus}
\title{Helper function for plot_diverse_homology_thresholds_core_orthologs}
\usage{
testCoreOrthoParamsGeneLocus(ortho_tables, param, species_order)
}
\description{
Helper function for plot_diverse_homology_thresholds_core_orthologs
}
