% Generated by roxygen2 (4.0.1): do not edit by hand
\name{transl}
\alias{transl}
\title{Function to translate an CDS sequence in string format to AA in string format}
\usage{
transl(sequence)
}
\arguments{
\item{sequence}{a character string specifying CDS sequence of interest.}
}
\value{
A character string specifying the corresponding amino acid sequence.
}
\description{
This function takes CDS sequence as string as input an returns the
corresponding amino acid sequence as string.
}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}

