% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_cds_to_protein.R
\name{translate_cds_to_protein}
\alias{translate_cds_to_protein}
\title{Translate coding sequences into amino acid sequences}
\usage{
translate_cds_to_protein(input_file, output_file, delete_corrupt_cds = FALSE)
}
\arguments{
\item{input_file}{file path to \code{fasta} file storing coding sequences (DNA).}

\item{output_file}{name or file path in which translated amino acid sequences (AA)
shall be stored as \code{fasta} file.}

\item{delete_corrupt_cds}{delete_corrupt_cds a logical value indicating whether sequences with corrupt base triplets should be removed from the input \code{file}. This is the case when the length of coding sequences cannot be divided by 3 and thus the coding sequence contains at least one corrupt base triplet.}
}
\description{
A helper function that takes a \code{fasta} file storing coding sequences
as input and translates these coding sequences into amino acid sequences
storing them as \code{fasta} output file.
}
\examples{
\dontrun{
# install.packages("biomartr")
# download coding sequences of Arabidopsis thaliana
Ath_file_path <- biomartr::getCDS(db = "refseq",
                     organism = "Arabidopsis thaliana",
                     gunzip = TRUE)
# translate coding sequences into amino acid sequences
translate_cds_to_protein(Ath_file_path, "Ath_aa_seqs.fasta")
}
}
\author{
Hajk-Georg Drost
}
