% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/io.R
\name{write.proteome}
\alias{write.proteome}
\title{Save a proteome in fasta format}
\usage{
write.proteome(proteome, file.name, nbchar = 80, open = "w",
  as.string = TRUE)
}
\arguments{
\item{proteome}{a data.table object storing amino acid sequences in the first column
and gene ids in the second column. See \code{\link{read.proteome}} for details.}

\item{file.name}{a character string specifying the name of the fasta output file.}

\item{nbchar}{number of characters per line.}

\item{as.string}{when set to TRUE sequences are in the form of strings instead of vectors of single characters.}

\item{mode}{to open the output file, use "w" to write into a new file, use "a" to append at the end of an already existing file.}
}
\description{
This function takes a proteome (in data.table notation) as returned by
\code{\link{read.proteome}} and saves it to your hard drive.
}
\examples{
\dontrun{

# read an example proteome
Ath.proteome <- read.proteome(system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'), format = "fasta")

# use the write.proteome function to store it on your hard drive
write.proteome(Ath.proteome, "test_proteome.fasta")

}
}
\author{
Hajk-Georg Drost
}

