% Generated by roxygen2 (4.0.1): do not edit by hand
\name{advanced_blast}
\alias{advanced_blast}
\title{Advanced interface function to BLAST+}
\usage{
advanced_blast(query_file, subject_file, format = "fasta",
  blast_algorithm = "blastp", path = NULL, blast_params = NULL,
  taxonomy = FALSE)
}
\arguments{
\item{query_file}{a character string specifying the path to the sequence file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the sequence file of interest (subject organism).}

\item{format}{a character string specifying the file format of the sequence file, e.g. "fasta", "gbk".}

\item{blast_algorithm}{a character string specifying the BLAST algorithm that shall be used, e.g. "blastp","blastn","tblastn","deltablast",... .}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{blast_params}{a character string listing the input paramters that shall be passed to the executing BLAST program. Default is \code{NULL}, implicating
that a set of default parameters is used when running BLAST.}

\item{taxonomy}{a logical value specifying whether the subject taxonomy shall be returned based on NCBI Taxonomy queries. Default is \code{NULL}.}
}
\value{
the \code{advanced_blast} function creates a folder named '_blast' and stores
the hit table returned by BLAST in this folder.
}
\description{
This function performs a BLAST+ search of a given set of sequences and a given set of parameters against a BLAST database.
}
\details{
Following BLAST programs and algorithms can be assigned to \code{blast_algorithm}:

"blastp", "blastn", "megablast","psi-blast", "phi-blast", "delta-blast", "blastx", "tblastn", "tblastx"
}
\note{
This function is also able to return the subject taxonomy.

For this purpose the NCBI taxdb must be installed on your system:

ftp://ftp.ncbi.nlm.nih.gov/blast/db/taxdb.tar.gz

In case you want to use the taxonomy option, set \code{taxonomy} = \code{TRUE}.
}
\examples{
\dontrun{


# performing a BLAST search using blastp (default)
advanced_blast(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
      subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
      blast_algorithm = "blastp", blast_params = "-evalue 1E-5 -num_threads 2")



}
}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}
\references{
Altschul, S.F., Gish, W., Miller, W., Myers, E.W. & Lipman, D.J. (1990) "Basic local alignment search tool." J. Mol. Biol. 215:403-410.

http://blast.ncbi.nlm.nih.gov/Blast.cgi?CMD=Web&PAGE_TYPE=BlastDocs&DOC_TYPE=ProgSelectionGuide

http://blast.ncbi.nlm.nih.gov/Blast.cgi
}

