% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blast}
\alias{blast}
\title{Interface function to BLAST+}
\usage{
blast(query_file, subject_file, eval = "1E-5", path = NULL,
  comp_cores = 1)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{comp_cores}{a numeric value specifying the number of cores that shall be
used to run BLAST searches}
}
\value{
A data.table storing the BLAST hit table returned by the BLAST program.
}
\description{
This function performs a blast+ search of a given set of protein sequences against a given database.
}
\examples{
\dontrun{
# performing a best hit BLAST search
blast(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
      subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'))

# in case you are working with a multicore machine, you can also run parallel
# BLAST computations using the comp_cores parameter: here with 2 cores
blast(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
      subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
      comp_cores = 2)
}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

