% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blast}
\alias{blast}
\title{Interface function to BLAST+}
\usage{
blast(query.dt, database, eval = "1E-5", input = "_blast/_blastinput.fasta",
  output = "_blast/_blastresult.csv", path = NULL, comp_cores = 1)
}
\arguments{
\item{query.dt}{a data.table returned as first list element of the set() function.}

\item{database}{a string specifying the path to the BLAST database, as returned
as second list element of the set() function.}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{input}{}

\item{output}{}

\item{path}{}

\item{comp_cores}{a numeric value specifying the number of cores that shall be
used to run BLAST searches}
}
\value{
A data.table storing the BLAST hit table returned by the BLAST program.
}
\description{
This function performs a blast+ search of a given set of protein sequences against a given database.
}
\examples{
\dontrun{
# performing a best hit BLAST search
blast( set("data/ortho_thal_cds.fasta")[[1]] , set("data/ortho_lyra_cds.fasta", makedb = TRUE)[[2]])
}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

