% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blast}
\alias{blast}
\title{Interface function to BLAST+}
\usage{
blast(queries, database, eval = "1E-5",
  blast_name = "_blast/_blastinput.fasta",
  blast_out = "_blast/_blastresult.csv", path = NULL)
}
\arguments{
\item{queries}{}

\item{database}{}

\item{eval}{}

\item{blast_name}{}

\item{blast_out}{}

\item{path}{}
}
\value{
A data.table storing the geneids of orthologous genes (best hit)
in the first column and the amino acid sequences in the second column.
}
\description{
This function performs a blast+ search of a given set of protein sequences against a given database.
}
\examples{
\dontrun{
# performing a best hit BLAST search
blast( set("ortho_lyra_cds.fasta") , set("ortho_thal_cds.fasta", makedb = TRUE))
}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

