% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blast_best}
\alias{blast_best}
\title{Function to perform a BLAST best hit search}
\usage{
blast_best(query_file, subject_file, path = NULL, comp_cores = 1)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{comp_cores}{a numeric value specifying the number of cores to be used for multicore BLAST computations.}
}
\value{
A data.table as returned by the blast() function, storing the geneids
of orthologous genes (best hit) in the first column and the amino acid sequences in the second column.
}
\description{
This function performs a blast+ search (best hit) of a given set of protein sequences against a given database.
}
\details{
Given a set of protein sequences A, a best hit blast search is being performed from A to database.
}
\examples{
\dontrun{
# performing gene orthology inference using the best hit (BH) method
blast_best(query_file = "data/ortho_thal_cds.fasta", subject_file = "data/ortho_lyra_cds.fasta")

# use multicore processing
blast_best(query_file = "data/ortho_thal_cds.fasta",
           subject_file = "data/ortho_lyra_cds.fasta",
           comp_cores = 2)
}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

