% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blast_best}
\alias{blast_best}
\title{Function to perform a BLAST best hit search}
\usage{
blast_best(query_file, subject_file, seq_type = "cds", format = "fasta",
  blast_algorithm = "blastp", eval = "1E-5", path = NULL,
  comp_cores = 1, blast_params = NULL, clean_folders = TRUE)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{seq_type}{a character string specifying the sequence type stored in the input file.
Options are are: "cds", "protein", or "dna". In case of "cds", sequence are translated to protein sequences,
in case of "dna", cds prediction is performed on the corresponding sequences which subsequently are
translated to protein sequences. Default is \code{seq_type} = "cds".}

\item{format}{a character string specifying the file format of the sequence file, e.g. "fasta", "gbk". Default is "fasta".}

\item{blast_algorithm}{a character string specifying the BLAST algorithm that shall be used, e.g. "blastp","blastn","tblastn",... .}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{comp_cores}{a numeric value specifying the number of cores to be used for multicore BLAST computations.}

\item{blast_params}{a character string listing the input paramters that shall be passed to the executing BLAST program. Default is \code{NULL}, implicating
that a set of default parameters is used when running BLAST.}

\item{clean_folders}{a logical value specifying whether the internal folder structure shall be deleted (cleaned) after
processing this function.}
}
\value{
A data.table as returned by the \code{blast} function, storing the geneids
of orthologous genes (best hit) in the first column and the amino acid sequences in the second column.
}
\description{
This function performs a blast+ search (best hit) of a given set of protein sequences against a given database.
}
\details{
Given a set of protein sequences A, a best hit blast search is being performed from A to database.
}
\examples{
\dontrun{

# performing gene orthology inference using the best hit (BH) method
blast_best(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
           subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'))

# performing gene orthology inference using the best hit (BH) method starting with protein sequences
blast_best(query_file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
           subject_file = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'),
           seq_type = "protein")

# use multicore processing
blast_best(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
           subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
           comp_cores = 2)

}
}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}
\seealso{
\code{\link{blast}}, \code{\link{blast_rec}}, \code{\link{advanced_blast}}, \code{\link{set_blast}}
}

