% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cds2aa}
\alias{cds2aa}
\title{Translate CDS file to Amino Acids file}
\usage{
cds2aa(dt)
}
\arguments{
\item{dt}{a \code{data.table} object storing cds sequences in a column named \code{seqs}.}
}
\description{
This function takes a \code{data.table} object as input and
translates the cds sequences stored as \code{seqs} column into the corresponding amino acid
sequence.
}

