% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_all_folders}
\alias{clean_all_folders}
\title{Function to delete the internal folder hierarchy}
\usage{
clean_all_folders(foldernames)
}
\arguments{
\item{foldernames}{a character vector storing the folder names that shall be deleted.}
}
\value{
a void function.
}
\description{
This function deletes all internal folders that have been created
during pipeline processing. Internally this function uses \code{\link{unlink}}
to delete all folders created by the pipline.
}
\details{
This function takes a vector storing the names of the folders
that shall be deleted, e.g.: \code{clean_all_folders( c("_alignment", "_blast_db", "_calculation") )}.
}
\examples{
\dontrun{

# in case internal folders exist, they are being removed
clean_all_folders(c("_alignment", "_blast_db", "_calculation"))

}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{divergence_stratigraphy}}
}

