% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clean_all_folders}
\alias{clean_all_folders}
\title{Function to delete the internal folder hierarchy}
\usage{
clean_all_folders()
}
\value{
Nothing is being returned.
}
\description{
This function deletes all internal folders that have been created
during pipeline processing. Internally this function uses \code{\link{unlink}}
to delete all folders created by the pipline.
}
\details{
Following folders are being removed: "_alignment", "_blast_db", and "_calculation".
}
\examples{
\dontrun{

# in case internal folders exist, they are being removed
clean_all_folders()

}
}
\author{
Hajk-Georg Drost
}
\seealso{
\code{\link{divergence_stratigraphy}}
}

