% Generated by roxygen2 (4.0.1): do not edit by hand
\name{codon_aln}
\alias{codon_aln}
\title{Function to get a codon alignment.}
\usage{
codon_aln(file.aln, file.nuc, file.out, format = "clustal", mode,
  get.aln = FALSE, path)
}
\arguments{
\item{file.aln}{a character string specifying the path to the file storing the protein alignment in CLUSTAL or FASTA format.}

\item{file.nuc}{a character string specifying the path to the file storing the coding sequences in multiple FASTA format.}

\item{file.out}{a character string specifying the path where the alignment should be stored.}

\item{format}{a character string specifying the file format used to store the codon alignment, e.g. "fasta", "clustal".}

\item{mode}{a character string specifying the program that should be used e.g. "pal2nal".}

\item{get.aln}{a logical value indicating whether the produced alignment should be returned.}

\item{path}{a character string specifiying the absolute path where pal2nal.pl can be found.}
}
\value{
if get.aln is TRUE an object of class alignment of the seqinr package.
}
\description{
This function takes a protein alignment and the corresponding coding sequences
both given as files and uses the method specified by mode to create the corresponding codon
alignment.
}
\details{
This function provides an interface between R and common codon alignment tools such as "PAL2NAL".
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

