% Generated by roxygen2 (4.0.1): do not edit by hand
\name{codon_aln}
\alias{codon_aln}
\title{Function to get a codon alignment.}
\usage{
codon_aln(file_aln, file_nuc, format = "clustal", tool, get_aln = FALSE,
  path)
}
\arguments{
\item{file_aln}{a character string specifying the path to the file storing the protein alignment in CLUSTAL or FASTA format.}

\item{file_nuc}{a character string specifying the path to the file storing the coding sequences in multiple FASTA format.}

\item{format}{a character string specifying the file format used to store the codon alignment, e.g. "fasta", "clustal".}

\item{tool}{a character string specifying the program that should be used e.g. "pal2nal".}

\item{get_aln}{a logical value indicating whether the produced alignment should be returned.}

\item{path}{a character string specifying the path to the codon alignment program (in case you don't use the default path).}
}
\value{
if get_aln is TRUE an object of class alignment of the seqinr package.
}
\description{
This function takes a protein alignment and the corresponding coding sequences
both given as files and uses the method specified by tool to create the corresponding codon
alignment.
}
\details{
This function provides an interface between R and common codon alignment tools such as "PAL2NAL".
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

