% Generated by roxygen2 (4.0.2): do not edit by hand
\name{compute_dnds}
\alias{compute_dnds}
\title{This function computes the dNdS value of one given pairwise alignment.}
\usage{
compute_dnds(complete_tbl, multialn_tool = "clustalw", multialn_path = NULL,
  multialn_params = NULL, codonaln_tool = "pal2nal",
  dnds_est.method = "YN", quiet = FALSE, comp_cores = 1)
}
\arguments{
\item{complete_tbl}{}

\item{multialn_tool}{}

\item{multialn_path}{}

\item{multialn_params}{}

\item{codonaln_tool}{}

\item{dnds_est.method}{}

\item{quiet}{}

\item{comp_cores}{}
}
\description{
This function takes a vector containing the query_id, subject_id,
query amino acid sequence, subject amino acid sequence, query CDS sequence, and subject CDS sequence
and then runs the following pipieline:

1) Multiple-Alignment of query amino acid sequence and subject amino acid sequence

2) Codon-Alignment of the amino acid alignment returned by 1) and query CDS sequence + subject CDS sequence

3) dNdS estimation of the codon alignment returned by 2)
}
\details{
This function takes the amino acid and CDS sequences two orthologous genes
and writes the corresponding amino acid and CDS sequences as fasta file into
the internal folder environment. The resulting fasta files (two files) store the
amino acid sequence of the query_id and subject_id (file one) and the CDS sequence of
the query_id and subject_id (file two). These fasta files are then used to pass through the following pipeline:

1) Multiple-Alignment of query amino acid sequence and subject amino acid sequence

2) Codon-Alignment of the amino acid alignment returned by 1) and query CDS sequence + subject CDS sequence

3) dNdS estimation of the codon alignment returned by 2)
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

