% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dNdS.R
\name{compute_dnds}
\alias{compute_dnds}
\title{This function computes the dNdS value of one given pairwise alignment}
\usage{
compute_dnds(complete_tbl, aa_aln_type = "multiple",
  aa_aln_tool = "clustalw", aa_aln_path = NULL, aa_aln_params = NULL,
  codon_aln_tool = "pal2nal", dnds_est.method = "YN",
  kaks_calc_path = NULL, quiet = FALSE, comp_cores = 1,
  clean_folders = FALSE)
}
\arguments{
\item{complete_tbl}{a data.table object storing the query_id, subject_id, query_cds (sequence),
subject_cds (sequence), query_aa (sequence), and subject_aa (sequence) of the organisms that shall be compared.}

\item{aa_aln_type}{a character string specifying the amino acid alignement type: \code{aa_aln_type} = "multiple" or \code{aa_aln_type} = "pairwise".
Default is \code{aa_aln_type} = "multiple".}

\item{aa_aln_tool}{a character string specifying the multiple alignment tool that shall be used for pairwise protein alignments.}

\item{aa_aln_path}{a character string specifying the path to the corresponding multiple alignment tool.}

\item{aa_aln_params}{a character string specifying additional parameters that shall be passed to the multiple alignment system call.}

\item{codon_aln_tool}{a character string specifying the codon alignment tool that shall be used for codon alignments. Default is \code{codon_aln_tool} = "pal2nal".}

\item{dnds_est.method}{a character string specifying the dNdS estimation method, e.g. "Comeron","Li", "YN", etc. See Details for all options.}

\item{kaks_calc_path}{a character string specifying the execution path to KaKs_Calculator. Default is \code{kaks_calc_path} = \code{NULL}
(meaning that KaKs_Calculator is stored and executable in your default \code{PATH}).}

\item{quiet}{a logical value specifying whether a successful interface call shall be printed out.}

\item{comp_cores}{a numeric value specifying the number of cores that shall be used to perform parallel computations on a multicore machine.}

\item{clean_folders}{a boolean value spefiying whether all internall folders storing the output of used programs
shall be removed. Default is \code{clean_folders} = \code{FALSE}.}
}
\description{
This function takes a vector containing the
query amino acid sequence, subject amino acid sequence, query CDS sequence, and subject CDS sequence
and then runs the following pipieline:

1) Multiple-Alignment of query amino acid sequence and subject amino acid sequence

2) Codon-Alignment of the amino acid alignment returned by 1) and query CDS sequence + subject CDS sequence

3) dNdS estimation of the codon alignment returned by 2)
}
\details{
This function takes the amino acid and CDS sequences two orthologous genes
and writes the corresponding amino acid and CDS sequences as fasta file into
the internal folder environment. The resulting fasta files (two files) store the
amino acid sequence of the query_id and subject_id (file one) and the CDS sequence of
the query_id and subject_id (file two). These fasta files are then used to pass through the following pipeline:

1) Multiple-Alignment or Pairwise-Alignment of query amino acid sequence and subject amino acid sequence

2) Codon-Alignment of the amino acid alignment returned by 1) and query CDS sequence + subject CDS sequence

3) dNdS estimation of the codon alignment returned by 2)
}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}
\references{
\url{http://www.r-bloggers.com/the-wonders-of-foreach/}
}
\seealso{
\code{\link{multi_aln}}, \code{\link{substitutionrate}}, \code{\link{dNdS}}
}

