% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DivergenceMap}
\alias{DivergenceMap}
\title{Function to sort dNdS values into divergence strata.}
\usage{
DivergenceMap(dNdS_tbl)
}
\arguments{
\item{dNdS_tbl}{a data.table object returned by \code{\link{dNdS}}.}
}
\value{
a data.table storing a standard divergence map.
}
\description{
This function takes a data.table returned by dNdS
and sorts the corresponding dNdS value into divergence strata (deciles).
}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}

