% Generated by roxygen2 (4.0.1): do not edit by hand
\name{dNdS}
\alias{dNdS}
\title{Function to calculate the synonymous vs nonsynonymous substitutionrate for two organisms.}
\usage{
dNdS(query_file, subject_file, blast_mode = "best hit", blast_path = NULL,
  multialn_tool = "clustalw", multialn_path = NULL,
  codonaln_tool = "pal2nal", codonaln_path = NULL,
  dnds_est.method = "Cameron", comp_cores = 1)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{blast_path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{multialn_path}{a character string specifying the path to the multiple alignment program (in case you don't use the default path).}

\item{codonaln_path}{a character string specifying the path to the codon alignment program (in case you don't use the default path).}

\item{dnds_est.method}{a character string specifying the dNdS estimation method, e.g. "Cameron","Li" .}

\item{comp_cores}{a numeric value specifying the number of cores that shall be}

\item{tool}{a character string specifying the program that should be used e.g. "clustalw".}
}
\value{
A data.table storing the dNdS values of the correspnding genes.
}
\description{
This function takes
}
\details{
This function...
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

