% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dNdS}
\alias{dNdS}
\title{Function to calculate the synonymous vs nonsynonymous substitutionrates for two organisms.}
\usage{
dNdS(query_file, subject_file, ortho_detection = "RBH", blast_path = NULL,
  multialn_tool = "clustalw", multialn_path = NULL,
  multialn_params = NULL, codonaln_tool = "pal2nal",
  dnds_est.method = "YN", comp_cores = 1, quiet = FALSE)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{ortho_detection}{a character string specifying the orthology inference method that shall be performed
to detect orthologous genes. Default is \code{ortho_detection} = "RBH" (BLAST reciprocal best hit).
Further methods are: "BH" (BLAST best hit), "RBH" (BLAST reciprocal best hit), "PO" (ProteinOrtho), "OrthoMCL, "IP" (InParanoid).}

\item{blast_path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{multialn_path}{a character string specifying the path to the multiple alignment program (in case you don't use the default path).}

\item{dnds_est.method}{a character string specifying the dNdS estimation method, e.g. "Comeron","Li" .}

\item{comp_cores}{a numeric value specifying the number of cores that shall be}

\item{tool}{a character string specifying the program that should be used e.g. "clustalw".}
}
\value{
A data.table storing the dNdS values of the correspnding genes.
}
\description{
This function takes the CDS files of two organisms of interest (query_file and subject_file)
and computes the dNdS estimation values for orthologous gene pairs between these organisms.
}
\details{
This function...
}
\examples{
\dontrun{

# get a dNdS table using:
# 1) reciprocal best hit for orthology inference (RBH)
# 2) clustalw for pairwise amino acid alignments
# 3) pal2nal for codon alignments
# 4) Yang, Z. and Nielsen, R. (2000) (YN) for dNdS estimation
# 5) single core processing 'comp_cores = 1'
dNdS(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
ortho_detection = "RBH",
multialn_tool = "clustalw", multialn_path = "/path/to/clustalw/",
codonaln_tool = "pal2nal", dnds_est.method = "YN", comp_cores = 1)

# The same result can be obtained using multicore processing using: comp_cores = 2 or 3 or more ...
dNdS(query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
subject_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
ortho_detection = "RBH",
multialn_tool = "clustalw", multialn_path = "/path/to/clustalw/",
codonaln_tool = "pal2nal", dnds_est.method = "YN", comp_cores = 2)

}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

