% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dNdS.R
\name{dNdS}
\alias{dNdS}
\title{Compute dNdS values for two organisms}
\usage{
dNdS(query_file, subject_file, seq_type = "cds", format = "fasta",
  ortho_detection = "RBH", delete_corrupt_cds = FALSE,
  cdd.path = NULL, blast_params = NULL, blast_path = NULL,
  eval = "1E-5", ortho_path = NULL, aa_aln_type = "pairwise",
  aa_aln_tool = "NW", aa_aln_path = NULL, aa_aln_params = NULL,
  codon_aln_tool = "pal2nal", kaks_calc_path = NULL,
  dnds_est.method = "Comeron", comp_cores = 1, quiet = TRUE,
  clean_folders = FALSE, print_citation = TRUE)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of interest (query organism).}

\item{subject_file}{a character string specifying the path to the CDS file of interest (subject organism).}

\item{seq_type}{a character string specifying the sequence type stored in the input file.Options are are: 
\itemize{
\item \code{seq_type = "cds"} (Default): sequence are translated to protein sequences
\item \code{seq_type = "protein"}: orthology inference is performed using protein sequences directly.
}}

\item{format}{a character string specifying the file format of the sequence file, e.g. \code{format = "fasta"}, \code{format = "gbk"}. See \code{\link{read.cds}},
\code{\link{read.genome}}, \code{\link{read.proteome}} for more details.}

\item{ortho_detection}{a character string specifying the orthology inference method that shall be performed to detect orthologous genes. Options are:
\itemize{
\item \code{ortho_detection ="BH"}: BLAST unidirectional best hit.
\item \code{ortho_detection = "RBH"}: BLAST reciprocal/bidirectional best hit (Default).
\item \code{ortho_detection = "Orthofinder2"}: single copy core orthologs between multiple species proteome comparisons.
}}

\item{delete_corrupt_cds}{a logical value indicating whether sequences with corrupt base triplets should be removed from the input \code{file}. This is the case when the length of coding sequences cannot be divided by 3 and thus the coding sequence contains at least one corrupt base triplet.}

\item{cdd.path}{path to the cdd database folder (specify when using \code{ortho_detection} = \code{"DELTA"}).}

\item{blast_params}{a character string specifying additional parameters that shall be passed to BLAST. Default is \code{blast_params} = \code{NULL}.}

\item{blast_path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{eval}{a numeric value specifying the E-Value cutoff for BLAST hit detection.}

\item{ortho_path}{a character string specifying the path to the orthology inference program such as \code{Orthofinder2}, etc. (in case you don't use the default path).}

\item{aa_aln_type}{a character string specifying the amino acid alignement type:
\itemize{ 
\item \code{aa_aln_type} = "multiple" (Default)
\item \code{aa_aln_type} = "pairwise"
}.}

\item{aa_aln_tool}{a character string specifying the program that should be used e.g. "clustalw".}

\item{aa_aln_path}{a character string specifying the path to the multiple alignment program (in case you don't use the default path).}

\item{aa_aln_params}{a character string specifying additional parameters that shall be passed to the selected alignment tool. Default is \code{aa_aln_params} = \code{NULL} 
(no addintional parameters are passed to the selected alignment tool).}

\item{codon_aln_tool}{a character string specifying the codon alignment tool that shall be used. Default is \code{codon_aln_tool} = \code{"pal2nal"}.
Right now only "pal2nal" can be selected as codon alignment tool.}

\item{kaks_calc_path}{a character string specifying the execution path to KaKs_Calculator. Default is \code{kaks_calc_path} = \code{NULL}
(meaning that KaKs_Calculator is stored and executable in your default \code{PATH}).}

\item{dnds_est.method}{a character string specifying the dNdS estimation method, e.g. "Comeron","Li", "YN", etc. See Details for all options.}

\item{comp_cores}{a numeric value specifying the number of cores that shall be used to perform parallel computations on a multicore machine.}

\item{quiet}{a logical value specifying whether the output of the corresponding alignment tool shall be printed out to the console.
Default is \code{quiet} = \code{FALSE}.}

\item{clean_folders}{a boolean value spefiying whether all internall folders storing the output of used programs
shall be removed. Default is \code{clean_folders} = \code{FALSE}.}

\item{print_citation}{a logical value indicating whether or not the citation message shall be printed.}
}
\value{
A data.table storing the dNdS values of the correspnding genes.
}
\description{
This function takes the CDS files of two organisms of interest (query_file and subject_file)
and computes the dNdS estimation values for orthologous gene pairs between these organisms.
}
\details{
The dN/dS ratio quantifies the mode and strength of selection acting on a pair of orthologous genes.
This selection pressure can be quantified by comparing synonymous substitution rates (dS) that are assumed to be neutral
with nonsynonymous substitution rates (dN), which are exposed to selection as they 
change the amino acid composition of a protein (Mugal et al., 2013 http://mbe.oxfordjournals.org/content/31/1/212).

The \pkg{orthologr} package provides the \code{\link{dNdS}} function to perform dNdS estimation on pairs of orthologous genes.
This function takes the CDS files of two organisms of interest (\code{query_file} and \code{subject_file}) 
and computes the dNdS estimation values for orthologous gene pairs between these organisms.

The following pipieline resembles the dNdS estimation process:

\itemize{
                
\item 1) Orthology Inference: e.g. BLAST reciprocal best hit (RBH)

\item 2) Pairwise sequence alignment: e.g. clustalw for pairwise amino acid sequence alignments

\item 3) Codon Alignment: e.g. pal2nal program

\item 4) dNdS estimation: e.g. Yang, Z. and Nielsen, R. (2000) \url{http://mbe.oxfordjournals.org/content/17/1/32.short}

}
Note: it is assumed that when using \code{dNdS()} all corresponding multiple sequence alignment programs you
want to use are already installed on your machine and are executable via either
the default execution \code{PATH} or you specifically define the location of the executable file
via the \code{aa_aln_path} or \code{blast_path} argument that can be passed to \code{dNdS()}.

The \code{dNdS()} function can be used choosing the folllowing options:

\itemize{
\item \code{ortho_detection} : 
   \itemize{  
   \item \code{"RBH"} (BLAST best reciprocal hit)
   \item \code{"BH"} (BLAST best hit)
   \item \code{"Orthofinder2"}
     }
     
\item \code{aa_aln_type} : 
 \itemize{
  \item \code{"multiple"}
  \item \code{"pairwise"}
 }

\item \code{aa_aln_tool} :
\itemize{
 \item \code{"clustalw"}
 \item \code{"t_coffee"}
 \item \code{"muscle"}
 \item \code{"clustalo"}
 \item \code{"mafft"}
 \item \code{"NW"} (in case \code{aa_aln_type = "pairwise"})
}

\item \code{codon_aln_tool} :
\itemize{
 \item \code{"pal2nal"}
 }
\item \code{dnds_est.method} : 
\itemize{
\item "Li" : Li's method (1993)
\item "Comeron" : Comeron's method (1995)
\item "NG": Nei, M. and Gojobori, T. (1986)
\item "LWL": Li, W.H., et al. (1985)
\item "LPB": Li, W.H. (1993) and Pamilo, P. and Bianchi, N.O. (1993)
\item "MLWL" (Modified LWL), MLPB (Modified LPB): Tzeng, Y.H., et al. (2004)
\item "YN": Yang, Z. and Nielsen, R. (2000)
\item "MYN" (Modified YN): Zhang, Z., et al. (2006)

}

}
}
\examples{
\dontrun{

# get a dNdS table using:
# 1) reciprocal best hit for orthology inference (RBH)
# 2) Needleman-Wunsch for pairwise amino acid alignments
# 3) pal2nal for codon alignments
# 4) Comeron for dNdS estimation
# 5) single core processing 'comp_cores = 1'
dNdS(query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
     subject_file    = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
     ortho_detection = "RBH", 
     aa_aln_type     = "pairwise",
     aa_aln_tool     = "NW", 
     codon_aln_tool  = "pal2nal", 
     dnds_est.method = "Comeron", 
     comp_cores      = 1 )


# running dNdS using the 'aa_aln_path' argument to specify the path to
# the corresponding alignment tool
dNdS(query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
     subject_file    = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
     ortho_detection = "RBH", 
     aa_aln_type     = "pairwise",
     aa_aln_tool     = "NW", 
     aa_aln_path     = "/path/to/clustalw/",
     codon_aln_tool  = "pal2nal", 
     dnds_est.method = "Comeron", 
     comp_cores      = 1 )

# The same result can be obtained using multicore processing using: comp_cores = 2 or 3 or more ...
dNdS(query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
     subject_file    = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
     ortho_detection = "RBH", 
     aa_aln_type     = "pairwise",
     aa_aln_tool     = "NW", 
     aa_aln_path     = "/path/to/clustalw/",
     codon_aln_tool  = "pal2nal", 
     dnds_est.method = "Comeron", 
     comp_cores      = 1 )

}
}
\references{
seqinr: \url{http://seqinr.r-forge.r-project.org/}

Zhang Z, Li J, Zhao XQ, Wang J, Wong GK, Yu J: KaKs Calculator: Calculating Ka and Ks through model selection and model averaging. Genomics Proteomics Bioinformatics 2006 , 4:259-263. 

KaKs_Calculator: \url{https://code.google.com/p/kaks-calculator/} [GNU GPL-3 license]

Paradis, E. (2012) Analysis of Phylogenetics and Evolution with R (Second Edition). New York: Springer.

Paradis, E., Claude, J. and Strimmer, K. (2004) APE: analyses of phylogenetics and evolution in R language. Bioinformatics, 20, 289-290.

More information on \pkg{ape} can be found at \url{http://ape-package.ird.fr/}.

Pages H, Aboyoun P, Gentleman R and DebRoy S. Biostrings: String objects representing biological sequences, and
matching algorithms. R package version 2.32.1.
}
\seealso{
\code{\link{divergence_stratigraphy}}, \code{\link{orthologs}}, \code{\link{substitutionrate}}, \code{\link{multi_aln}}, \code{\link{codon_aln}}, \code{\link{blast_best}},
\code{\link{blast_rec}}, \code{\link{read.cds}}
}
\author{
Hajk-Georg Drost
}
