% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_core_set.R
\name{filter_core_set}
\alias{filter_core_set}
\title{Helper function to extract a core set of orthologous gene loci}
\usage{
filter_core_set(x, order_species)
}
\description{
Helper function to extract a core set of orthologous gene loci
}
