% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_core_set_lnc.R
\name{filter_core_set_lnc}
\alias{filter_core_set_lnc}
\title{Helper function to extract a core set of orthologous lncRNAs}
\usage{
filter_core_set_lnc(x, order_species)
}
\description{
Helper function to extract a core set of orthologous lncRNAs
}
