% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_core_set_lnc.R
\name{filter_core_set_lnc}
\alias{filter_core_set_lnc}
\title{Helper function to extract a core set of orthologous lncRNAs}
\usage{
filter_core_set_lnc(x, order_species)
}
\arguments{
\item{x}{input data in \code{data.frame} or \code{tibble} format.}

\item{order_species}{a character vector containing the scientific names of the organisms of interest
ordered according to their phylogenetic distance to their reference species.}
}
\description{
Helper function to extract a core set of orthologous lncRNAs.
}
