% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnc_map_core_orthologs.R
\name{lnc_map_core_orthologs}
\alias{lnc_map_core_orthologs}
\title{Retrieve a core set of orthologous lncRNAs from the pairwise lncRNA orthologs map}
\usage{
lnc_map_core_orthologs(lnc_map, species_order)
}
\arguments{
\item{lnc_map}{a \code{lnc_map} that was generated with \code{\link{map_generator_lnc}}.}

\item{species_order}{a character string specifying species names listed in the order of phylogenetic/taxonomic distance from the query species.
The species names must match with the species names present in the \code{lnc_map}.}
}
\description{
Given a lnc_map table generated with \code{\link{map_generator_lnc}},
this function will determine a core set of lncRNA orthologs that are shared between all species.
}
\author{
Hajk-Georg Drost
}
