% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.generator.R
\name{map.generator}
\alias{map.generator}
\title{Generate dNdS Maps Between a Query Organism and Multiple Subject Organisms}
\usage{
map.generator(query_file, subjects.folder, output.folder, eval = "1E-5",
  ortho_detection = "RBH", comp_cores = 1, ds.values = FALSE,
  subject.id = TRUE, progress.bar = TRUE, sep = ";", ...)
}
\arguments{
\item{query_file}{a character string specifying the path to the CDS file of the query organism.}

\item{subjects.folder}{a character string specifying the path to the folder where CDS files of the subject organisms are stored.}

\item{output.folder}{a character string specifying the path to the folder where output dNdS/DS Maps should be stored stored.}

\item{eval}{a character string specifying the e-value for BLAST based Orthology Inference that is performed
in the process of dNdS computations. Please use the scientific notation.}

\item{ortho_detection}{a character string specifying the Orthology Inference method that shall be used to perform
dNdS computations. Possible options are: \code{ortho_detection} = \code{"BH"} (BLAST best hit),
\code{ortho_detection} = \code{"RBH"} (BLAST best reciprocal hit), etc.}

\item{comp_cores}{number of computing cores that shall be used to perform parallelized computations.}

\item{ds.values}{a logical value indicating whether divegrence stratum values (ds values) or dNdS values should be returned by \code{\link{divergence_stratigraphy}}. Default is \code{ds.values = FALSE}.}

\item{subject.id}{a logical value indicating whether \code{query_id} AND \code{subject_id} should be returned.}

\item{progress.bar}{should a progress bar be shown. Default is \code{progress.bar = TRUE}.}

\item{sep}{a file separator that is used to store maps as csv file.}

\item{...}{additional parameters that shall be passed to  \code{\link{divergence_stratigraphy}}.}
}
\description{
This function allows you to compute dNdS or DS Maps between a query organism
and a set subject organisms stored in the same folder. The corresponding dNdS/DS Maps are then stored in an output folder.
}
\details{
Given a query organism and a set of subject organsisms that are stored in the same folder,
this function crawls through all subject organsism and as a first step computes the pairwise
dNdS or DS Maps between query and subject organsim and as a second step stores the corresponding Map
in an output folder.
}
\examples{
\dontrun{

map.generator(
   query_file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
   subjects.folder = system.file('seqs/map_gen_example', package = 'orthologr'),
   output.folder = getwd(),
   quiet = TRUE
)

}
}
\author{
Hajk-Georg Drost
}

