% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.generator.lnc.R
\name{map_generator_lnc}
\alias{map_generator_lnc}
\title{Infer orthologous lncRNAs between multiple species}
\usage{
map_generator_lnc(query_file, subjects_folder, output_folder,
  eval = "1E-5", ortho_detection = "RBH", max.target.seqs = 10000,
  min_qry_coverage_hsp = 30, min_qry_perc_identity = 30,
  comp_cores = 1, progress_bar = TRUE, sep = ";", ...)
}
\arguments{
\item{query_file}{a character string specifying the path to the lncRNAs file of the query organism in \code{fasta} format.}

\item{subjects_folder}{a character string specifying the path to the folder where lncRNAs files in \code{fasta} format of the subject organisms are stored.}

\item{output_folder}{a character string specifying the path to the folder where output orthologous tables should be stored.}

\item{eval}{a character string specifying the e-value for BLAST based orthology inference. Please use the scientific notation.}

\item{ortho_detection}{a character string specifying the Orthology Inference method that shall be used to perform
dNdS computations. Possible options are: 
\itemize{
 \item \code{ortho_detection = "BH"}: BLAST best unidirectional hit
 \item \code{ortho_detection = "RBH"}: BLAST best reciprocal hit
}}

\item{max.target.seqs}{a numeric value specifying the number of aligned sequences to keep.
Please be aware that \code{max.target.seqs} selects best hits based on the database entry and not by the best e-value. See details here: https://academic.oup.com/bioinformatics/advance-article/doi/10.1093/bioinformatics/bty833/5106166 .}

\item{min_qry_coverage_hsp}{minimum \code{qcovhsp} (= query coverage of the HSP) of an orthologous hit (a value between 1 and 100).}

\item{min_qry_perc_identity}{minimum \code{perc_identity} (= percent sequence identity between query and selected HSP) of an orthologous hit (a value between 1 and 100).}

\item{comp_cores}{number of computing cores that shall be used to perform parallelized computations.}

\item{progress_bar}{should a progress bar be shown. Default is \code{progress_bar = TRUE}.}

\item{sep}{a file separator that is used to store maps as csv file.}

\item{...}{additional parameters that shall be passed to  \code{\link{dNdS}}.}
}
\description{
Inference of orthologous lncRNAs between multiple species is performed via pairwise BLAST (reciprocal) best hit comparisons.
 The corresponding orthologous tables are then stored in an output folder.
}
\details{
Given a query organism and a set of subject organsisms that are stored in the same folder,
this function crawls through all subject organsism and infers the lncRNA homologs in
pairwise species comparisons.
}
\examples{
\dontrun{
map_generator_lnc(
   query_file      = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
   subjects_folder = system.file('seqs/map_gen_example', package = 'orthologr'),
   output_folder   = getwd(),
   quiet           = TRUE,
   comp_cores      = 1
)
}
}
\author{
Hajk-Georg Drost
}
