% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multi_aln}
\alias{multi_aln}
\title{Function to get a multiple alignment.}
\usage{
multi_aln(file, file.out, mode, get.aln = "FALSE")
}
\arguments{
\item{file}{a character string specifying the path to the file storing the sequences in FASTA format.}

\item{file.out}{a character string specifying the path where the alignment should be stored.}

\item{mode}{a character string specifying the program that should be used e.g. "clustalw".}

\item{get.aln}{a logical value indicating whether the produced alignment should be returned.}
}
\value{
if get.aln is TRUE an object of class alignment of the seqinr package.
}
\description{
This function takes a multiple FASTA file containing DNA or aminoacid sequences
to be aligned and proceeds the method defined by mode to align them.
}
\details{
This function provides an interface between R and common multiple alignment programs
such as "clustalw", "tcoffee", "muscle", and "clustalo".
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

