% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multi_aln}
\alias{multi_aln}
\title{An interface function to common multiple alignment tools.}
\usage{
multi_aln(file, tool, get_aln = "FALSE", path = NULL)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the sequences in FASTA format.}

\item{tool}{a character string specifying the program that should be used: "clustalw", "tcoffee", "muscle", and "clustalo".}

\item{get_aln}{a logical value indicating whether the produced alignment should be returned.}

\item{path}{a character string specifying the path to the multiple alignment program (in case you don't use the default path).}
}
\value{
If get_aln is TRUE an object of class alignment of the seqinr package.
}
\description{
This function takes a multiple FASTA file containing DNA or amino acid sequences
that shall be aligned and computes a multiple alignment using a defined multiple alignment tool.
}
\details{
This function provides an interface between R and common multiple alignment programs
such as "clustalw", "tcoffee", "muscle", and "clustalo".

CLUSTALW :

Different operating systems perform different execution calls to the clustalw program:

MacOS: 'clustalw2', Linux: 'clustalw', Windows: 'clustalw2.exe'

In case you use the default path to the clustalw program, depending on your operating system,
the following calls to clustalw should work properly on your system:

MacOS: system("clustalw2 -help"), Linux: system("clustalw -help"), Windows: system("clustalw2.exe -help")

In case these procedures don't work properly, please use the \code{path} argument
to specify the 'clustalw' execution path on your system:


MacOS: system("path/to/clustalw/clustalw2 -help"), Linux: system("path/to/clustalw/clustalw -help"), Windows: system("path/to/clustalw/clustalw2.exe -help")
}
\examples{
\dontrun{

# CLUSTALW Example:

# in case the default execution path of clustalw runs properly on your system
multi_aln(system.file('seqs/aa_seqs.fasta', package = 'orthologr'),
          tool = "clustalw", get_aln = TRUE)

# in case the default execution path of clustalw is not set within the default path
multi_aln(system.file('seqs/aa_seqs.fasta', package = 'orthologr'),
          tool = "clustalw", get_aln = TRUE, path = "path/to/clustalw/")


}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

