% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multi_aln}
\alias{multi_aln}
\title{Function to get a multiple alignment.}
\usage{
multi_aln(file, tool, get_aln = "FALSE", path = NULL)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the sequences in FASTA format.}

\item{tool}{a character string specifying the program that should be used e.g. "clustalw".}

\item{get_aln}{a logical value indicating whether the produced alignment should be returned.}

\item{path}{a character string specifying the path to the multiple alignment program (in case you don't use the default path).}
}
\value{
If get_aln is TRUE an object of class alignment of the seqinr package.
}
\description{
This function takes a multiple FASTA file containing DNA or aminoacid sequences
to be aligned and proceeds the method defined by tool to align them.
}
\details{
This function provides an interface between R and common multiple alignment programs
such as "clustalw", "tcoffee", "muscle", and "clustalo".
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

