% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthofinder2.R
\name{orthofinder2}
\alias{orthofinder2}
\title{Interface function to Orthofinder2}
\usage{
orthofinder2(proteome_folder, comp_cores = 1)
}
\arguments{
\item{proteome_folder}{file path to a folder storing the proteome sequences of the species for which orthology inference shall be performed.}

\item{comp_cores}{number of cores that shall be used for parallel processing. Default is \code{cores = 1}.}
}
\description{
Run Orthofinder2 from R.
}
\examples{
\dontrun{
# download proteome files
biomartr::getProteome(db = "refseq", organism = "Arabidopsis thaliana", path = "of_proteomes")
biomartr::getProteome(db = "refseq", organism = "Arabidopsis lyrata", path = "of_proteomes")
biomartr::getProteome(db = "refseq", organism = "Capsella rubella", path = "of_proteomes")
biomartr::getProteome(db = "refseq", organism = "Solanum lycopersicum", path = "of_proteomes")

} 
}
\author{
Hajk-Georg Drost
}
