% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthofinder2.R
\name{orthofinder2}
\alias{orthofinder2}
\title{Interface function to Orthofinder2}
\usage{
orthofinder2(proteome_folder, comp_cores = 1)
}
\arguments{
\item{proteome_folder}{file path to a folder storing the proteome sequences of the species for which orthology inference shall be performed.}

\item{comp_cores}{number of cores that shall be used for parallel processing. Default is \code{cores = 1}.}
}
\description{
Run Orthofinder2 from R.
}
\examples{
\dontrun{
# specify species names
orgs <- c("Arabidopsis thaliana", "Arabidopsis lyrata", 
          "Capsella rubella", "Solanum lycopersicum")
# download proteome files for all species          
biomartr::getProteomeSet(db = "refseq", organisms = orgs, path = "of_proteomes")
# download annotation files for all species          
biomartr::getGFFSet(db = "refseq", organisms = orgs, path = "of_gff")
# select longest splice variant per gene locus
retrieve_longest_isoforms_all(proteome_folder = "of_proteomes", 
                              annotation_folder = "of_gff",
                              annotation_format = "gff", 
                              output_folder = "of_proteomes_longest_sv")
# run orthofinder2 to infer ortho groups for the specified species
orthofinder2(proteome_folder = "of_proteomes_longest_sv", comp_cores = 4)
} 
}
\author{
Hajk-Georg Drost
}
