% Generated by roxygen2 (4.0.2): do not edit by hand
\name{orthologs}
\alias{orthologs}
\title{Interface function to perform orthology inference for a set of organisms}
\usage{
orthologs(organisms, format = "fasta", ortho_detection = "RBH")
}
\arguments{
\item{organisms}{}

\item{format}{}

\item{ortho_detection}{}
}
\value{
a data.table storing ...
}
\description{
This function takes nucleotide or protein sequences for a set of organisms
and performs orthology inference to detect orthologous genes within the given organisms
based on selected orthology inference programs.
}
\details{
This function...
}
\author{
Hajk-Georg Drost
}

