% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairwise_aln}
\alias{pairwise_aln}
\title{An interface function to common pairwise alignment tools.}
\usage{
pairwise_aln(file, tool = "NW", seqtype, get_aln = FALSE,
  pairwise_aln_name = NULL, path = NULL, quiet = FALSE)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the sequences in FASTA format.}

\item{tool}{a character string specifying the program/algorithm that should be used: "NW".}

\item{get_aln}{a logical value indicating whether the produced alignment should be returned.}

\item{path}{a character string specifying the path to the pairwise alignment program (in case you don't use the default path).}

\item{paiwise_aln_name}{a character string specifying the name of the stored alignment file.
Default is \code{pairwise_aln_name} = \code{NULL} denoting a default name: 'toolname_seqtype.aln' .}
}
\value{
In case the argument \code{get_aln} is set \code{TRUE}, an object of class alignment of the seqinr package is returned.
}
\description{
This function takes a FASTA file containing two DNA or amino acid sequences
that shall be aligned and computes a paiwise alignment using a defined alignment method.
}
\details{
This function provides an interface between R and common pairwise alignment programs.
}
\examples{
\dontrun{

# Needleman-Wunsch Example:

# in case Biostrings works properly
pairwise_aln(system.file('seqs/aa_seqs.fasta', package = 'orthologr'),
             tool = "NW", get_aln = TRUE, seqtype="AA")

}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

