% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/plot_diverse_homology_thresholds_core_orthologs.R
\name{plot_diverse_homology_thresholds_core_orthologs}
\alias{plot_diverse_homology_thresholds_core_orthologs}
\title{Diverse line plots visualizing the number of core orthologs within a ortho table generated with \code{\link{generate_ortholog_tables_all}}
based on different sets of homology thresholds.}
\usage{
plot_diverse_homology_thresholds_core_orthologs(ortho_tables,
  species_order, type = "both")
}
\arguments{
\item{ortho_tables}{a \code{ortho tables} that was generated with \code{\link{generate_ortholog_tables_all}}.}

\item{species_order}{a character string specifying species names listed in the order of phylogenetic/taxonomic distance from the query species.
The species names must match with the species names present in the \code{ortho_tables}.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label.}

\item{title}{plot title.}
}
\description{
Given a ortho table generated with \code{\link{generate_ortholog_tables_all}},
this function will visualize the number of core orthologs 
using a diverse range of homology thresholds to enable an analytical decision for chosing homology thresholds to define orthologous genes.
}
\author{
Hajk-Georg Drost
}
