% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.genome}
\alias{read.genome}
\title{Function to read a genome of a given organism}
\usage{
read.genome(file, format = "fasta", ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the genome.}

\item{format}{a character string specifying the file format used to store the genome, e.g. "fasta", "gb".}

\item{...}{additional arguments that are used by the seqinr::read.fasta() function.}
}
\description{
This function reads an organism specific genome stored in a defined file format.
}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}

