% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_core_orthologs.R
\name{retrieve_core_orthologs}
\alias{retrieve_core_orthologs}
\title{Retrieve a core set of orthologous gene loci from several pairwise ortholog tables}
\usage{
retrieve_core_orthologs(ortho_tables, species_order)
}
\arguments{
\item{ortho_tables}{a \code{ortho tables} that was generated with \code{\link{generate_ortholog_tables_all}}.}

\item{species_order}{a character string specifying species names listed in the order of phylogenetic/taxonomic distance from the query species.
The species names must match with the species names present in the \code{ortho_tables}.}
}
\description{
Given a ortho table generated with \code{\link{generate_ortholog_tables_all}},
this function will determine a core set of orthologs that are shared between all species.
}
\author{
Hajk-Georg Drost
}
