% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set_blast}
\alias{set_blast}
\title{Function preparing the parameters and databases for subsequent BLAST searches}
\usage{
set_blast(file, format = "fasta", makedb = FALSE, path = NULL, ...)
}
\arguments{
\item{file}{a character string specifying the path to the file storing the cd.}

\item{format}{a character string specifying the file format used to store the genome, e.g. "fasta", "gbk".}

\item{makedb}{TRUE or FALSE whether a database should be created or not.}

\item{path}{a character string specifying the path to the BLAST program (in case you don't use the default path).}

\item{...}{additional arguments that are used by the seqinr::read.fasta() function.}
}
\value{
A list with [[1]] a data.table storing the gene id in the first column and
the corresponding dna and aminoacid sequence as string in the second and third column
respectively and [[2]] the name of the database created.
}
\description{
This function reads a cds fasta file using read.cds(), translates each
sequence into the corresponding aminoacid sequence and is able to create a blast
database from that.
}
\examples{
\dontrun{
 # running the set function to see an example output
 head(set_blast(file = system.file('ortho_thal_cds.fasta', package = 'orthologr'))[[1]] , 2)
}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

