% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ssearch}
\alias{ssearch}
\title{Interface to the commandline tool ssearch36.}
\usage{
ssearch(file, library_file, options = "", lib_format = "fasta",
  seq_type = "protein", path = NULL, parse_output_to = NULL)
}
\arguments{
\item{file}{a character string specifying the path to the sequence file.}

\item{library_file}{a character string specifying the path to the library file.}

\item{options}{a character string specifying additional options for ssearch}

\item{lib_format}{a character string specifying the file type of the library e.g. fasta.}

\item{seq_type}{a character string specifying the type of the sequences}

\item{path}{a character string specifying the path to ssearch if not included in
the global environment}

\item{parse_output_to}{a charcater string specifying a file where the output of
ssearch is parsed to}
}
\description{
This function is a wrapper for the tool ssearch that performs
global alignment for each sequence in the inputfile with each sequence in the
library and returns the alignment with the best score.
}
\details{
"Perform a rigorous Global/Global (ssearch) [...] alignment between
a protein sequence and another protein sequence or a protein database [or
a DNA sequence to a DNA sequence database ]". ssearch is "using the Smith-Waterman
algorithm. ssearch36 uses SSE2 acceleration, and is only 2 - 5X slower than
fasta36
}
\examples{
\dontrun{
# performing ssearch on two protein files
ssearch( file = system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'),
          library_file = system.file('seqs/ortho_lyra_aa.fasta', package = 'orthologr'))

# performing ssearch on two cds files parsing the output
ssearch( file = system.file('seqs/ortho_thal_cds.fasta', package = 'orthologr'),
          library_file = system.file('seqs/ortho_lyra_cds.fasta', package = 'orthologr'),
          seqtype="cds", parse_output_to = "ssearch.out")

}
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}
\references{
ssearch is part of the fasta36 package that can be found here
http://fasta.bioch.virginia.edu/fasta_www2/fasta_down.shtml
}

