% Generated by roxygen2 (4.0.1): do not edit by hand
\name{substitutionrate}
\alias{substitutionrate}
\title{Function to calculate the synonymous vs nonsynonymous substitutionrate for a codon alignment.}
\usage{
substitutionrate(file, est.method, format = "fasta")
}
\arguments{
\item{file}{a character string specifying the path to a codon alignment file}

\item{est.method}{a character string specifying the dNdS estimation method, e.g. "Cameron","Li" "mase", "clustal", "phylip", "fasta" , "msf".
Note, that when using "Cameron" as dNdS estimation method, the program 'gestimator' is used to compute the
corresponding dNdS values from a given alignment. The program 'gestimator' can only read "fasta" files,
hence it is important to use format = "fasta" when choosing est.method = "Cameron".}

\item{format}{a character string specifying the file format in which the alignment is stored:}
}
\value{
some value
}
\description{
This function takes
}
\details{
This function ...
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

