% Generated by roxygen2 (4.0.1): do not edit by hand
\name{substitutionrate}
\alias{substitutionrate}
\title{Function to calculate the synonymous vs nonsynonymous substitutionrate for a codon alignment.}
\usage{
substitutionrate(file, tool, format = "fasta", path = NULL)
}
\arguments{
\item{file}{a character string specifying the path to a codon alignment file}

\item{format}{a character string specifying the format a used in read.alignment (for Li)}
}
\value{
some value
}
\description{
This function takes
}
\details{
This function ...
}
\author{
Sarah Scharfenberg and Hajk-Georg Drost
}

