% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCoreOrthoParamsSplicceVariant.R
\name{testCoreOrthoParamsSpliceVariant}
\alias{testCoreOrthoParamsSpliceVariant}
\title{Helper function for plot_diverse_homology_thresholds_core_orthologs 2}
\usage{
testCoreOrthoParamsSpliceVariant(ortho_tables, param, species_order,
  n_core_species)
}
\description{
Helper function for plot_diverse_homology_thresholds_core_orthologs 2
}
