% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transl.R
\name{transl}
\alias{transl}
\title{Translate DNA to Amino Acids}
\usage{
transl(sequence)
}
\arguments{
\item{sequence}{a character string specifying CDS sequence of interest.}
}
\value{
A character string specifying the corresponding amino acid sequence.
}
\description{
This function takes CDS sequence as string as input an returns the
corresponding amino acid sequence as string.
}
\examples{

# an example DNA sequence
DNA <- c("ACCGGTTTAAAGGCGTTA")

# translating DNA to a protein sequence
transl(DNA)

}
\author{
Hajk-Georg Drost and Sarah Scharfenberg
}
\references{
\code{\link[seqinr]{translate}}
}

