% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.proteome.R
\name{write.proteome}
\alias{write.proteome}
\title{Save a proteome in fasta format}
\usage{
write.proteome(proteome, file.name, nbchar = 80, open = "w",
  as.string = TRUE)
}
\arguments{
\item{proteome}{a data.table object storing amino acid sequences in the first column
and gene ids in the second column. See \code{\link{read.proteome}} for details.}

\item{file.name}{a character string specifying the name of the fasta output file.}

\item{nbchar}{number of characters per line.}

\item{open}{mode to open the output file, you can choose from "w" to write into a new file, or "a" to append at the end of an already existing file.}

\item{as.string}{when set to TRUE sequences are in the form of strings instead of vectors of single characters.}
}
\description{
This function takes a proteome (in data.table notation) as returned by 
\code{\link{read.proteome}} and saves it to your hard drive.
}
\examples{

\dontrun{

# read an example proteome
Ath.proteome <- read.proteome(
                      system.file('seqs/ortho_thal_aa.fasta', package = 'orthologr'), 
                      format = "fasta")

# use the write.proteome function to store it on your hard drive
write.proteome(Ath.proteome, "test_proteome.fasta")

}

}
\author{
Hajk-Georg Drost
}
